### 2021-02-21 #110 @whedon
[PRE REVIEW]: Neuromatch Academy: a 3-week, online summer school in computational neuroscience
**Submitting author:** @thartbm (<a href="http://orcid.org/0000-0002-0045-3769">Bernard Marius &#39;t Hart</a>)
**Repository:** <a href="https://github.com/NeuromatchAcademy/course-content" target ="_blank">https://github.com/NeuromatchAcademy/course-content</a>
**Version:** 1

@thartbm if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
