### 2021-02-22 #110 @thartbm
The line mentioned there has a URL in the affiliation (it's an event that has changing physical locations, but a constant web address). I'm guessing that parts of the URL are read to have meaning in yaml (like a colon), can I prevent that somehow?
