### 2021-02-22 #110 @labarba
I don't think so. You can try things around and then [Check that your paper compiles](https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles)
