+++ bibliography.bib
@@ -6,7 +6,7 @@
 	abstract = {Neuromatch Academy designed and ran a fully online 3-week Computational Neuroscience summer school for 1757 students with 191 teaching assistants working in virtual inverted (or flipped) classrooms and on small group projects. Fourteen languages, active community management, and low cost allowed for an unprecedented level of inclusivity and universal accessibility.},
 	urldate = {2021-02-22},
 	journal = {arXiv:2012.08973 [q-bio]},
-	author = {van Viegen, Tara and Akrami, Athena and Bonnen, Kate and DeWitt, Eric and Hyafil, Alexandre and Ledmyr, Helena and Lindsay, Grace W. and Mineault, Patrick and Murray, John D. and Pitkow, Xaq and Puce, Aina and Sedigh-Sarvestani, Madineh and Stringer, Carsen and Achakulvisut, Titipat and Alikarami, Elnaz and Atay, Melvin Selim and Batty, Eleanor and Erlich, Jeffrey C. and Galbraith, Byron V. and Guo, Yueqi and Juavinett, Ashley L. and Krause, Matthew R. and Li, Songting and Pachitariu, Marius and Straley, Elizabeth and Valeriani, Davide and Vaughan, Emma and Vaziri-Pashkam, Maryam and Waskom, Michael L. and Blohm, Gunnar and Kording, Konrad and Schrater, Paul and Wyble, Brad and Escola, Sean and Peters, Megan A. K.},
+	author = "{van Viegen}, Tara and Akrami, Athena and Bonnen, Kate and DeWitt, Eric and Hyafil, Alexandre and Ledmyr, Helena and Lindsay, Grace W. and Mineault, Patrick and Murray, John D. and Pitkow, Xaq and Puce, Aina and Sedigh-Sarvestani, Madineh and Stringer, Carsen and Achakulvisut, Titipat and Alikarami, Elnaz and Atay, Melvin Selim and Batty, Eleanor and Erlich, Jeffrey C. and Galbraith, Byron V. and Guo, Yueqi and Juavinett, Ashley L. and Krause, Matthew R. and Li, Songting and Pachitariu, Marius and Straley, Elizabeth and Valeriani, Davide and Vaughan, Emma and Vaziri-Pashkam, Maryam and Waskom, Michael L. and Blohm, Gunnar and Kording, Konrad and Schrater, Paul and Wyble, Brad and Escola, Sean and Peters, Megan A. K.",
 	month = dec,
 	year = {2020},
 	note = {arXiv: 2012.08973},
