### 2021-02-22 #110 @thartbm
Oh, interesting. I only submitted once that compiled. Since I did change the file a bit since first submitting (including replacing one of the colons from the affiliation on line 494 with a comma), I just checked and it had no problems. I'm not sure what to change now.

There are some additional things to work on regarding author names. If you look at the one citation in the paper, I had to indicate what the last name of the first author is, as it contains a space, like my own name. I don't know how to indicate this in the yaml header though so if we leave it as as, this is probably going to enter some people into databases with part of their last name as a first name. There are also two people with their name in uppercase, and this is correct, but might get "fixed" down the line.
