### 2021-02-22 b24e7fb
fixed "NA" in two sets of initials
