### 2021-02-22 cc45b54
consistent author contributions table + shortened
