+++ paper.md
@@ -63,7 +63,7 @@ authors:
  - name: John S Butler
    orcid: 0000-0001-5567-8634
    affiliation: 18
- - name: ccnss CCNSS
+ - name: CCNSS
    affiliation: 19
  - name: Brandon Caie
    affiliation: 14
@@ -822,18 +822,18 @@ For self-study or mixing the NMA course content with other content, we provide a
 | Team                           | Members                                                      |
 | ------------------------------ | ------------------------------------------------------------ |
 | Curriculum Design & Management | Akrami A; Blohm G; Bonnen K; DeWitt EEJ; Escola S; Hyafil A; Kording K; Kumar A; Macke JH; Mineault PJ; Murray JD; Peters MAK; Pitkow X; Schrater P; Stringer C; Valton V; Wallisch P; Wyble B; d'Oleire Uquillas F; van Viegen T |
-| Group Projects                 | Akrami A, Alicea B, Butler JS, DeWitt EEJ, Gupta PK, LC RAY, Lee J, Mohammadi N, Murray JD, Pachitariu M, Parent J, Pitkow X, Rodriguez-Cruces R, Saxena S, Schrater P, Sharma S, Steinmetz NA, Straley E, Strandquist G, Valton V, Waskom M, Zhang L |
-| Day Coordinators               | W0D1: Brigham M, W1D2: Blohm G, W1D3: Hyafil A, W1D4: Macke JH, W1D5: Murray JD, W2D1: Kording K, W2D2: Brunton BW, W2D3: Wallisch P, W2D4: Pitkow X, W2D5: DeWitt EEJ, W3D1: Kumar A, W3D2: Kumar A, W3D3: Kording K, W3D4: Stringer C, W3D5: DeWitt EEJ |
-| Intro & Outro Lecturers        | W0D1: Brigham M, W1D1: Blohm G, Peters MAK, W1D2: Blohm G, Peters MAK, W1D3: Drugowitsch J, Wei K, W1D4: Park IM, Savin C, W2D1: Kording K, Schrater P, Pitkow X, W2D2: Shea-Brown E, Shenoy KV, W2D3: Escola S, Shanechi MM, W2D4: Shadmehr R, Smith M, W3D1: Bhalla U, Poirazi P, W3D2: Miller KD, W3D3: Kording K, Kass RE, Triantafillou S, W3D4: Johnson M, Mur M, W3D5: Rolnick D |
-| Micro-Lecturers                | W0D1: Brigham M, W1D1: Kording K, W1D2: Blohm G, Schrater P, W1D3: Wu A, W1D4: Wu A, W1D5: Cayco Gajic NA, W2D1: Kording K, Valton V, W2D2: Brunton BW, W2D3: Fei Y, W2D4: Pitkow X, W2D5: DeWitt EEJ, Mattar MG, W3D1: Naud R, W3D2: Gjorgjieva J, W3D3: Kording K, Liu T, W3D4: Stringer C, Mohsenzadeh Y, W3D5: Brigham M |
[…]
