### 2021-02-22 #110 @thartbm
This is correct: Patrick J. Mineault does not have an affiliation, and that is as it should be. Is there some way I should indicate this?
