### 2021-02-22 #110 @labarba
Yes, you can [indicate it as "Independent Researcher,"](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) or something similar.
