### 2021-02-22 fd4ab54
Patrick's affiliation: "Independent Researcher"
