### 2021-02-22 #110 @thartbm
Alright, I asked Patrick and he's ok with "Independent Researcher". So I've added it, and the PDF compiles but seems to just print the yaml header across ~13 pages.
