### 2021-02-22 #110 @moorepants
FYI, see the conversation here for reference to the size and scope: https://github.com/openjournals/jose/issues/52
