### 2021-02-22 #110 @labarba
@thartbm — thank you for linking to the pre-submission enquiry. I have read that now, and got caught up with the discussion about authorship and size of the work submitted.

I have concerns about the author list.

Authorship guidelines vary from field to field but:

> The minimum requirements for authorship, common to all definitions, are 1) substantial contribution to the work and 2) accountability for the work that was done and its presentation in a publication. […]
> Acknowledgements may be used to denote contributions to the work that do not meet the criteria of authorship such as, supporting the study, general mentoring, collecting data, acting as study coordinator, and other related activities

[COPE Guidelines PDF, p.4](https://publicationethics.org/files/COPE_DD_A4_Authorship_SEPT19_SCREEN_AW.pdf)
