### 2021-02-22 #110 @koerding
As a co-founder of NMA and participant in the run-up to this paper I can confirm that we made sure that all authors substantially contributed (and that we spent weeks as a group making that sure for every person). I can also confirm that they all are and feel accountable (and that we explicitly asked each participant).
