### 2021-02-22 #110 @thartbm
As reviewers, these 6 people from your list seem to have a related background and knowledge of Python:

Ariel Rokem (arokem)
Teon Brooks (teonbrooks)
Enrico Glerean (eglerean)
Damien Mannion (djmannion)
Manjaru Narayan (mnarayan)
Thomas Donoghue (TomDonoghue)
