+++ paper.md
@@ -312,7 +312,7 @@ authors:
    affiliation: 24
  - name: Xaq Pitkow
    orcid: 0000-0001-6376-329X
-   affiliation: 43
+   affiliation: 37,43
  - name: Panayiota Poirazi
    orcid: 0000-0001-6152-595X
    affiliation: 22
