### 2021-05-13 e631286
Use markdown table and emoji to mimic an admonition box
