### 2021-05-13 f7f04f2
Update readme (#550)
* Added under construction warning

* Make link nicer

* Use markdown table and emoji to mimic an admonition box

Co-authored-by: Michael Waskom <mwaskom@users.noreply.github.com>
