### 2021-05-13 3f2b901
Prevent admonition from wrapping in an ugly way
