+++ README.md
@@ -1,7 +1,8 @@
 # NeuroMatch Academy (NMA) Computational Neuroscience syllabus
 
-| :warning: This repository is under construction for 2021. See the [NMA website](https://academy.neuromatch.io/nma2020/course-materials) for materials from the 2020 course. :warning: |
-| --- |
+| :warning: This repository is under construction for the 2021 edition of the course :warning:
+| :---: |
+| See the [NMA website](https://academy.neuromatch.io/nma2020/course-materials) for links to archival materials from the 2020 edition |
 
 **Objectives**: Introduce traditional and emerging computational neuroscience tools, their complementarity, and what they can tell us about the brain. A main focus is on modeling choices, model creation, model evaluation and understanding how they relate to biological questions.
 
