### 2021-05-14 614c67e
Fix the regex that replaces branch names to support hyphens
