### 2021-05-24 3e13750
initial script to generate book toc
