### 2021-05-24 e69e238
master reference of all course materials, used to generate jupyter book
