+++ tutorials/materials.yml
@@ -1,0 +1,295 @@
+- day: W0D1
+  name: Python Workshop 1
+  tutorials: 1
+- day: W0D2
+  name: Python Workshop 2
+  tutorials: 1
+- day: W1D1
+  intro: https://www.youtube.com/watch?v=KxldhMR5PxA
+  name: Model Types
+  outro: https://www.youtube.com/watch?v=KZQXfQL1SH4
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18ObAiSOZ42YBwOQIKNvspeI
+  qa:
+  - https://www.youtube.com/watch?v=FwsKmunwzUA
+  - https://www.youtube.com/watch?v=h63bKG0-h_w
+  - https://www.youtube.com/watch?v=8h0X9lxY8Bs
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/rbx2a/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/6dxwe/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/9hkg2/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/2esh5/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/5vj73/?direct%26mode=render%26action=download%26mode=render
+    title: TA
+  tutorials: 3
+- day: W1D2
+  intro: https://www.youtube.com/watch?v=8pz_NH5_Zy4
+  name: Modeling Practice
+  outro: https://www.youtube.com/watch?v=Il8zOmCMFAA
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18Nc7rjBNO99bZQyuTY0TAcE
+  qa:
+  - https://www.youtube.com/watch?v=j8xsEOg8F_Q
+  - https://www.youtube.com/watch?v=VAM-suHZPSM
+  - https://www.youtube.com/watch?v=ha72Q_0e3mA
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/kmwus/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/kygfn/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/agrp6/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/2esh5/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/v5emn/?direct%26mode=render%26action=download%26mode=render
+    title: TA
+  tutorials: 2
+- day: W1D3
+  intro: https://www.youtube.com/watch?v=yTQNqYCvoZ0
+  name: Model Fitting
+  outro: https://www.youtube.com/watch?v=cUAUL3efqpY
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18MqTMnc_mMiuawnczEa8xm9
+  qa:
+  - https://www.youtube.com/watch?v=lYLa6zgqByw
+  - https://www.youtube.com/watch?v=LtR2rWiF7Mk
+  - https://www.youtube.com/watch?v=JhRqg3BAXlI
[…]
