+++ tutorials/W3D4_DeepLearning1/solutions/W3D4_Tutorial3_Solution_805a9425.py
@@ -1,22 +1,0 @@
-def RDM(resp):
-  """Compute the representational dissimilarity matrix (RDM)
-
-  Args:
-    resp (ndarray): S x N matrix with population responses to
-      each stimulus in each row
-
-  Returns:
-    ndarray: S x S representational dissimilarity matrix
-  """
-
-  # z-score responses to each stimulus
-  zresp = zscore(resp, axis=1)
-
-  # Compute RDM
-  RDM = 1 - (zresp @ zresp.T) / zresp.shape[1]
-
-  return RDM
-
-rdm_dict = {label: RDM(resp) for label, resp in resp_dict.items()}
-with plt.xkcd():
-  plot_multiple_rdm(rdm_dict)
