+++ tutorials/W3D4_DeepLearning1/solutions/W3D4_Tutorial2_Solution_185efdad.py
@@ -1,8 +1,0 @@
-
-"""
-We can think of all neurons being composed of the same shaped peak but at
-different/multiple locations for each neuron. However, it appears that different
-neurons have peaks that are shaped differently, or even the same neuron has two
-peaks with different shapes. Therefore we may want multiple different convolutional
-filters.
-""";
