+++ tutorials/W3D4_DeepLearning1/solutions/W3D4_Tutorial3_Solution_42c2b339.py
@@ -1,37 +1,0 @@
-def plot_rdm_rows(ori_list, rdm_dict, rdm_oris):
-  """Plot the dissimilarity of response to each stimulus with response to one
-  specific stimulus
-
-  Args:
-    ori_list (list of float): plot dissimilarity with response to stimulus with
-      orientations closest to each value in this list
-    rdm_dict (dict): RDM's from which to extract dissimilarities
-    rdm_oris (np.ndarray): orientations corresponding to each row/column of RDMs
-    in rdm_dict
-
-  """
-  n_col = len(ori_list)
-  f, axs = plt.subplots(1, n_col, figsize=(4 * n_col, 4), sharey=True)
-
-  # Get index of orientation closest to ori_plot
-  for ax, ori_plot in zip(axs, ori_list):
-    iori = np.argmin(np.abs(rdm_oris - ori_plot))
-
-    # Plot dissimilarity curves in each RDM
-    for label, rdm in rdm_dict.items():
-      ax.plot(rdm_oris, rdm[iori, :], label=label)
-
-    # Draw vertical line at stimulus we are plotting dissimilarity w.r.t.
-    ax.axvline(rdm_oris[iori], color=".7", zorder=-1)
-
-    # Label axes
-    ax.set_title(f'Dissimilarity with response\nto {ori_plot: .0f}$^o$ stimulus')
-    ax.set_xlabel('Stimulus orientation ($^o$)')
-
-  axs[0].set_ylabel('Dissimilarity')
-  axs[-1].legend(loc="upper left", bbox_to_anchor=(1, 1))
-
-
-ori_list = [-75, -25, 25, 75]
-with plt.xkcd():
-  plot_rdm_rows(ori_list, rdm_dict, ori)
