+++ tutorials/W3D4_DeepLearning1/solutions/W3D4_Tutorial1_Solution_8f827dbe.py
@@ -1,66 +1,0 @@
-
-def train(net, loss_fn, train_data, train_labels, n_iter=50, learning_rate=1e-4):
-  """Run gradient descent to opimize parameters of a given network
-
-  Args:
-    net (nn.Module): PyTorch network whose parameters to optimize
-    loss_fn: built-in PyTorch loss function to minimize
-    train_data (torch.Tensor): n_train x n_neurons tensor with neural
-      responses to train on
-    train_labels (torch.Tensor): n_train x 1 tensor with orientations of the
-      stimuli corresponding to each row of train_data, in radians
-    n_iter (int): number of iterations of gradient descent to run
-    learning_rate (float): learning rate to use for gradient descent
-
-  """
-
-  # Initialize PyTorch SGD optimizer
-  optimizer = optim.SGD(net.parameters(), lr=learning_rate)
-
-  # Placeholder to save the loss at each iteration
-  track_loss = []
-
-  # Loop over epochs (cf. appendix)
-  for i in range(n_iter):
-
-    # Evaluate loss using loss_fn
-    out = net(train_data)  # compute network output from inputs in train_data
-    loss = loss_fn(out, train_labels)  # evaluate loss function
-
-    # Compute gradients
-    optimizer.zero_grad()
-    loss.backward()
-
-    # Update weights
-    optimizer.step()
-
-    # Store current value of loss
-    track_loss.append(loss.item())  # .item() needed to transform the tensor output of loss_fn to a scalar
-
-    # Track progress
-    if (i + 1) % (n_iter // 5) == 0:
-      print(f'iteration {i + 1}/{n_iter} | loss: {loss.item():.3f}')
-
-  return track_loss
-
-# Set random seeds for reproducibility
-np.random.seed(1)
-torch.manual_seed(1)
-
-# Initialize network
-net = DeepNetReLU(n_neurons, 20)
-
-# Initialize built-in PyTorch MSE loss function
-loss_fn = nn.MSELoss()
-
-# Run GD on data
-train_loss = train(net, loss_fn, resp_train, stimuli_train)
-
-# Plot the training loss over iterations of GD
-with plt.xkcd():
-  plt.plot(train_loss)
-  plt.xlim([0, None])
-  plt.ylim([0, None])
-  plt.xlabel('iterations of gradient descent')
-  plt.ylabel('mean squared error')
-  plt.show()

[…]
