+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial4_Solution_07dba4f3.py
@@ -1,17 +1,0 @@
-
-"""
-Discussion:
-
-1. Yes. Because
-a. area under the STDP curve is negative
-b. synapses are set to their maximum possible weight.
-c. presynaptic neurons have a variance in their firing rates because of Poisson nature
-
-2. g_E will fluctuate even when synapses do not show any STDP because of the Poisson
-nature but with STDP, there will be bigger fluctuations as synaptic weight will
-vary on spike by spike basis.
-
-3. Individual synaptic weight will always fluctuate but the synaptic weight
-distribution will reach a steady state.
-
-""";
