+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial3_Solution_bedc4823.py
@@ -1,9 +1,0 @@
-
-"""
-Discussion:
-
-Increasing the input rate, we decrease the synaptic efficacy, i.e., the synaptic
-conductance decreases. This is the case for both Poisson or a regular spiking input.
-In case of regular spiking, the synaptic conductance reaches a steady state. This
-will not happen in the case of Poisson type spikes.
-""";
