+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial1_Solution_7a9c48e8.py
@@ -1,60 +1,0 @@
-def run_LIF(pars, Iinj, stop=False):
-  """
-  Simulate the LIF dynamics with external input current
-
-  Args:
-    pars       : parameter dictionary
-    Iinj       : input current [pA]. The injected current here can be a value
-                 or an array
-    stop       : boolean. If True, use a current pulse
-
-  Returns:
-    rec_v      : membrane potential
-    rec_sp     : spike times
-  """
-  # Set parameters
-  V_th, V_reset = pars['V_th'], pars['V_reset']
-  tau_m, g_L = pars['tau_m'], pars['g_L']
-  V_init, E_L = pars['V_init'], pars['E_L']
-  dt, range_t = pars['dt'], pars['range_t']
-  Lt = range_t.size
-  tref = pars['tref']
-
-  # Initialize voltage and current
-  v = np.zeros(Lt)
-  v[0] = V_init
-  Iinj = Iinj * np.ones(Lt)
-  if stop:  # set end of current to 0 if current pulse
-    Iinj[:int(len(Iinj) / 2) - 1000] = 0
-    Iinj[int(len(Iinj) / 2) + 1000:] = 0
-  tr = 0.  # the count for refractory duration
-
-  # Simulate the LIF dynamics
-  rec_spikes = []  # record spike times
-  for it in range(Lt - 1):
-    if tr > 0:  # check for refractoriness
-      v[it] = V_reset
-      tr = tr - 1
-    elif v[it] >= V_th:  # reset voltage and record spike event
-      rec_spikes.append(it)
-      v[it] = V_reset
-      tr = tref / dt
-
-    # calculate the increment of the membrane potential
-    dv = (-(v[it] - E_L) + Iinj[it] / g_L) * (dt / tau_m)
-
-    # update the membrane potential
-    v[it + 1] = v[it] + dv
-
-  rec_spikes = np.array(rec_spikes) * dt
-
-  return v, rec_spikes
-
-
-pars = default_pars(T=500)
-# Uncomment below to test your function
-v, sp = run_LIF(pars, Iinj=100, stop=True)
-with plt.xkcd():
-  plot_volt_trace(pars, v, sp)
-  plt.ylim([-80, -60])
-  plt.show()
