+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial1_Solution_fc553243.py
@@ -1,12 +1,0 @@
-
-"""
-Discussion:
-
-In a limiting case, when the time constant of the OU process is very long and the
-input current is almost flat, we expect the firing rate to decrease and neuron
-will spike more regularly. So as the OU process time constant increases, we expect
-firing rate and CV_ISI to decrease, if all other parameters are kept constant. We
-can also relate the OU process time constant to the membrane time constant as the
-neuron membrane does the same operation. This way we can link to the very first
-interactive demo.
-""";
