+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial3_Solution_789f00e0.py
@@ -1,28 +1,0 @@
-pars = default_pars(T=1000.)
-# Add parameters
-pars['gE_bar'] = 2.4    # [nS]
-pars['VE'] = 0.         # [mV] excitatory reversal potential
-pars['tau_syn_E'] = 2.  # [ms]
-pars['gI_bar'] = 2.4    # [nS]
-pars['VI'] = -80.       # [mV] inhibitory reversal potential
-pars['tau_syn_I'] = 5.  # [ms]
-
-# generate presynaptic spike trains
-pre_spike_train_ex = Poisson_generator(pars, rate=10, n=80)
-pre_spike_train_in = Poisson_generator(pars, rate=10, n=20)
-
-# simulate conductance-based LIF model
-v, rec_spikes, gE, gI = run_LIF_cond(pars, 0, pre_spike_train_ex,
-                                     pre_spike_train_in)
-dt, range_t = pars['dt'], pars['range_t']
-if rec_spikes.size:
-  sp_num = (rec_spikes / dt).astype(int) - 1
-  v[sp_num] = 10   # draw nicer spikes
-
-# Change the threshold
-pars['V_th'] = 1e3
-# Calculate FMP
-v_fmp, _, _, _ = run_LIF_cond(pars, 0, pre_spike_train_ex, pre_spike_train_in)
-
-with plt.xkcd():
-  my_illus_LIFSYN(pars, v_fmp, v)
