+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial4_Solution_1ba13cde.py
@@ -1,39 +1,0 @@
-def Delta_W(pars, A_plus, A_minus, tau_stdp):
-  """
-  Plot STDP biphasic exponential decaying function
-
-  Args:
-    pars       : parameter dictionary
-    A_plus     : (float) maxmimum amount of synaptic modification
-                 which occurs when the timing difference between pre- and
-                 post-synaptic spikes is positive
-    A_plus     : (float) maxmimum amount of synaptic modification
-                 which occurs when the timing difference between pre- and
-                 post-synaptic spikes is negative
-    tau_stdp   : the ranges of pre-to-postsynaptic interspike intervals
-                 over which synaptic strengthening or weakening occurs
-
-  Returns:
-    dW         : instantaneous change in weights
-  """
-
-  # STDP change
-  dW = np.zeros(len(time_diff))
-  # Calculate dW for LTP
-  dW[time_diff <= 0] = A_plus * np.exp(time_diff[time_diff <= 0] / tau_stdp)
-  # Calculate dW for LTD
-  dW[time_diff > 0] = -A_minus * np.exp(-time_diff[time_diff > 0] / tau_stdp)
-
-  return dW
-
-
-pars = default_pars_STDP()
-# Get parameters
-A_plus, A_minus, tau_stdp = pars['A_plus'], pars['A_minus'], pars['tau_stdp']
-# pre_spike time - post_spike time
-time_diff = np.linspace(-5 * tau_stdp, 5 * tau_stdp, 50)
-
-# Uncomment to test your function
-dW = Delta_W(pars, A_plus, A_minus, tau_stdp)
-with plt.xkcd():
-  mySTDP_plot(A_plus, A_minus, tau_stdp, time_diff, dW)
