+++ tutorials/W3D1_RealNeurons/solutions/W3D1_Tutorial2_Solution_e5eaac3e.py
@@ -1,45 +1,0 @@
-def corr_coeff_pairs(pars, rate, c, trials, bins):
-  """
-  Calculate the correlation coefficient of two spike trains, for different
-  realizations
-
-  Args:
-      pars   : parameter dictionary
-      rate   : rate of poisson inputs
-      c      : correlation coefficient ~ [0, 1]
-      trials  : number of realizations
-      bins   : vector with bins for time discretization
-
-  Returns:
-    r12      : correlation coefficient of a pair of inputs
-  """
-
-  r12 = np.zeros(n_trials)
-
-  for i in range(n_trials):
-    # Generate correlated Poisson inputs
-    sp1, sp2 = generate_corr_Poisson(pars, poi_rate, c, myseed=2020+i)
-
-    # Bin the spike times of the first input
-    sp1_count, _ = np.histogram(sp1, bins=bins)
-
-    # Bin the spike times of the second input
-    sp2_count, _ = np.histogram(sp2, bins=bins)
-
-    # Calculate the correlation coefficient
-    r12[i] = my_CC(sp1_count, sp2_count)
-
-  return r12
-
-
-poi_rate = 20.
-c = 0.2  # set true correlation
-pars = default_pars(T=10000)
-# bin the spike time
-bin_size = 20  # [ms]
-my_bin = np.arange(0, pars['T'], bin_size)
-n_trials = 100  # 100 realizations
-
-r12 = corr_coeff_pairs(pars, rate=poi_rate, c=c, trials=n_trials, bins=my_bin)
-print(f'True corr coe = {c:.3f}')
-print(f'Simu corr coe = {r12.mean():.3f}')
