### 2021-05-24 ba29de1
remove old dynamic networks folder
