+++ tutorials/W3D2_DynamicNetworks/solutions/W3D2_Tutorial1_Solution_5a95a98e.py
@@ -1,11 +1,0 @@
-
-"""
-Discussion:
-
-1) As the F-I curve is bounded between zero and one, the system doesn't explode.
-The f-curve guarantees this property
-
-2) One way to increase the maximum response is to change the f-I curve. For
-example, the ReLU is an unbounded function, and thus will increase the overall maximal
-response of the network.
-""";
