+++ tutorials/W3D2_DynamicNetworks/solutions/W3D2_Tutorial2_Solution_5a629797.py
@@ -1,15 +1,0 @@
-def EIderivs(rE, rI,
-             tau_E, a_E, theta_E, wEE, wEI, I_ext_E,
-             tau_I, a_I, theta_I, wIE, wII, I_ext_I,
-             **other_pars):
-  """Time derivatives for E/I variables (dE/dt, dI/dt)."""
-  # Compute the derivative of rE
-  drEdt = (-rE + F(wEE * rE - wEI * rI + I_ext_E, a_E, theta_E)) / tau_E
-  # Compute the derivative of rI
-  drIdt = (-rI + F(wIE * rE - wII * rI + I_ext_I, a_I, theta_I)) / tau_I
-
-  return drEdt, drIdt
-
-
-with plt.xkcd():
-  plot_complete_analysis(default_pars())
