+++ tutorials/W3D2_DynamicNetworks/solutions/W3D2_Tutorial1_Solution_4d2de6a0.py
@@ -1,12 +1,0 @@
-"""
-Discussion:
-
-To better appreciate what is happening here, you should go back to the previous
-interactive demo. Set the w = 5 and I_ext = 0.5.
-
-You will find that there are three fixed points of the system for these values of
-w and I_ext. Now, choose the initial value in this demo and see in which direction
-the system output moves. When r_init is in the vicinity of the leftmost fixed points
-it moves towards the left most fixed point. When r_init is in the vicinity of the
-rightmost fixed points it moves towards the rightmost fixed point.
-""";
