+++ tutorials/W3D2_DynamicNetworks/solutions/W3D2_Tutorial2_Solution_2366ea57.py
@@ -1,46 +1,0 @@
-def get_E_nullcline(rE, a_E, theta_E, wEE, wEI, I_ext_E, **other_pars):
-  """
-  Solve for rI along the rE from drE/dt = 0.
-
-  Args:
-    rE    : response of excitatory population
-    a_E, theta_E, wEE, wEI, I_ext_E : Wilson-Cowan excitatory parameters
-    Other parameters are ignored
-
-  Returns:
-    rI    : values of inhibitory population along the nullcline on the rE
-  """
-  # calculate rI for E nullclines on rI
-  rI = 1 / wEI * (wEE * rE - F_inv(rE, a_E, theta_E) + I_ext_E)
-
-  return rI
-
-
-def get_I_nullcline(rI, a_I, theta_I, wIE, wII, I_ext_I, **other_pars):
-  """
-  Solve for E along the rI from dI/dt = 0.
-
-  Args:
-    rI    : response of inhibitory population
-    a_I, theta_I, wIE, wII, I_ext_I : Wilson-Cowan inhibitory parameters
-    Other parameters are ignored
-
-  Returns:
-    rE    : values of the excitatory population along the nullcline on the rI
-  """
-  # calculate rE for I nullclines on rI
-  rE = 1 / wIE * (wII * rI + F_inv(rI, a_I, theta_I) - I_ext_I)
-
-  return rE
-
-
-pars = default_pars()
-Exc_null_rE = np.linspace(-0.01, 0.96, 100)
-Inh_null_rI = np.linspace(-.01, 0.8, 100)
-
-# Uncomment these lines to test your functions
-Exc_null_rI = get_E_nullcline(Exc_null_rE, **pars)
-Inh_null_rE = get_I_nullcline(Inh_null_rI, **pars)
-
-with plt.xkcd():
-  plot_nullclines(Exc_null_rE, Exc_null_rI, Inh_null_rE, Inh_null_rI)
