### 2021-05-24 02fc9f5
remove old bayes folder
