+++ tutorials/W2D1_BayesianStatistics/solutions/W2D1_Tutorial2_Solution_4d414185.py
@@ -1,14 +1,0 @@
-"""
-The mixture of Gaussian prior creates some interesting behaviour:
-  1. We observe multiple modes (i.e. peaks) in our posterior
-  (the common and independent causes).
-  2. The mode of the posterior jumps between stimulus locations. These
-  correspond to the participant switching from the independent to the common
-  parts (i.e. causes) of the prior.
-
-A similar discontinuity (ie. 'jump') in the posterior mode would happen in the
-case of cue combination illusion with the puppet and puppeteer voice.
-The illusion that the puppet is generating the speech would break-down when the
-voice stimulus is presented too far away from the visual input (the puppet's
-location).
-"""
