+++ tutorials/W2D5_ReinforcementLearning/solutions/W2D5_Tutorial3_Solution_725e14ff.py
@@ -1,29 +1,0 @@
-def q_learning(state, action, reward, next_state, value, params):
-  """Q-learning: updates the value function and returns it.
-
-  Args:
-    state (int): the current state identifier
-    action (int): the action taken
-    reward (float): the reward received
-    next_state (int): the transitioned to state identifier
-    value (ndarray): current value function of shape (n_states, n_actions)
-    params (dict): a dictionary containing the default parameters
-
-  Returns:
-    ndarray: the updated value function of shape (n_states, n_actions)
-  """
-  # Q-value of current state-action pair
-  q = value[state, action]
-
-  # write an expression for finding the maximum Q-value at the current state
-  if next_state is None:
-    max_next_q = 0
-  else:
-    max_next_q = np.max(value[next_state])
-
-  # write the expression to compute the TD error
-  td_error = reward + params['gamma'] * max_next_q - q
-  # write the expression that updates the Q-value for the state-action pair
-  value[state, action] = q + params['alpha'] * td_error
-
-  return value
