+++ tutorials/W2D5_ReinforcementLearning/solutions/W2D5_Tutorial2_Solution_457741c9.py
@@ -1,4 +1,0 @@
-"""
-When epsilon is zero, the agent always chooses the currently best option; it
-becomes greedy. When epsilon is 1, the agent chooses randomly.
-""";
