+++ tutorials/W2D5_ReinforcementLearning/solutions/W2D5_Tutorial1_Solution_dea47c05.py
@@ -1,4 +1,0 @@
-"""
-The TD trace now takes on negative values because the reward delivered is
-sometimes larger than the expected reward and sometimes smaller.
-""";
