+++ tutorials/W2D4_DynamicNetworks/solutions/W2D4_Tutorial1_Solution_c5280901.py
@@ -1,0 +1,24 @@
+def compute_drdt(r, I_ext, w, a, theta, tau, **other_pars):
+  """Given parameters, compute dr/dt as a function of r.
+
+  Args:
+    r (1D array) : Average firing rate of the excitatory population
+    I_ext, w, a, theta, tau (numbers): Simulation parameters to use
+    other_pars : Other simulation parameters are unused by this function
+
+  Returns
+    drdt function for each value of r
+  """
+  # Calculate drdt
+  drdt = (-r + F(w * r + I_ext, a, theta)) / tau
+
+  return drdt
+
+
+# Define a vector of r values and the simulation parameters
+r = np.linspace(0, 1, 1000)
+pars = default_pars_single(I_ext=0.5, w=5)
+
+drdt = compute_drdt(r, **pars)
+with plt.xkcd():
+  plot_dr_r(r, drdt)
