+++ tutorials/W3D4_ReinforcementLearning/solutions/W3D4_Tutorial2_Solution_457741c9.py
@@ -1,0 +1,4 @@
+"""
+When epsilon is zero, the agent always chooses the currently best option; it
+becomes greedy. When epsilon is 1, the agent chooses randomly.
+""";
