+++ tutorials/W2D4_DynamicNetworks/solutions/W2D4_Tutorial1_Solution_20486792.py
@@ -1,0 +1,13 @@
+
+"""
+Discussion:
+
+The fixed points of the single excitatory neuron population are determined by both
+recurrent connections w and external input I_ext. In a previous interactive demo
+we saw how the system showed two different steady-states when w = 0. But when w
+doe not equal 0, for some range of w the system shows three fixed points (the middle
+one being unstable) and the steady state depends on the initial conditions (i.e.
+r at time zero.).
+
+More on this will be explained in the next section.
+""";
