### 2021-05-25 a86c23d
Schedule rearrangement (#556)
* change tutorial file names

* remove old DL folder

* rename real neuron tutorials

* remove old real neurons  folder

* update real neurons headers

* update deep learning headers

* update dynamic networks file names

* remove old dynamic networks folder

* update dynamic networks headers

* rename bayes tutorials

* remove old bayes folder

* update bayes headers

* rename decision making files

* update decision making headers

* rename optimal control files

* update optimal control headers

* rename RL files

* remove old RL folder

* update RL headers

* 
[…]
