+++ tutorials/W3D3_OptimalControl/solutions/W3D3_Tutorial2_Solution_5ecbeb0c.py
@@ -1,0 +1,4 @@
+"""
+You should have seen that the Kalman filter generally estimates the latent state
+accurately, even with fairly high noise levels, except when C=0.
+"""
