### 2021-05-26 93fabb1
Install CI tools from centralized repo
