+++ .github/workflows/notebook-pr.yaml
@@ -35,9 +35,10 @@ jobs:
       - name: Install CI tools
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
         run: |
-          wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/main.tar.gz
-          tar -xzf main.tar.gz
-          mv nmaci-main/scripts/ ci/
+          BRANCH=`python -c 'import os, re; m = re.search(r"nmaci:([\w-]+)", os.environ["COMMIT_MESSAGE"]); print("main" if m is None else m.group(1))'`
+          wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/$BRANCH.tar.gz
+          tar -xzf $BRANCH.tar.gz
+          mv nmaci-$BRANCH/scripts/ ci/
 
       - name: Install dependencies
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
