### 2021-05-27 321a6dd
Instal the CI tools for the tutorial README PR too
