### 2021-05-27 5d06388
Prevent nmaci files from ending up in the sendback manifest
