+++ .github/workflows/tutorial-readme-pr.yaml
@@ -31,6 +31,8 @@ jobs:
           wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/$BRANCH.tar.gz
           tar -xzf $BRANCH.tar.gz
           mv nmaci-$BRANCH/scripts/ ci/
+          rm -r nmaci-main
+          echo ci/ >> .gitignore
 
       - name: Update READMEs
         run: python ci/generate_tutorial_readmes.py
