### 2021-05-27 40b0f1b
Install CI deps from nmaci file
