+++ .github/workflows/notebook-pr.yaml
@@ -38,8 +38,9 @@ jobs:
           BRANCH=`python -c 'import os, re; m = re.search(r"nmaci:([\w-]+)", os.environ["COMMIT_MESSAGE"]); print("main" if m is None else m.group(1))'`
           wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/$BRANCH.tar.gz
           tar -xzf $BRANCH.tar.gz
+          pip install -r nmaci-$BRANCH/requirements.txt
           mv nmaci-$BRANCH/scripts/ ci/
-          rm -r nmaci-main
+          rm -r nmaci-$BRANCH
           echo ci/ >> .gitignore
 
       - name: Install dependencies
@@ -47,7 +48,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip wheel
           pip install -r requirements.txt
-          pip install nbconvert pillow flake8 fuzzywuzzy[speedup]
 
       - name: Install XKCD fonts
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
