+++ .github/workflows/tutorial-readme-pr.yaml
@@ -31,7 +31,7 @@ jobs:
           wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/$BRANCH.tar.gz
           tar -xzf $BRANCH.tar.gz
           mv nmaci-$BRANCH/scripts/ ci/
-          rm -r nmaci-main
+          rm -r nmaci-$BRANCH
           echo ci/ >> .gitignore
 
       - name: Update READMEs
