+++ tutorials/W2D4_DynamicNetworks/solutions/W2D4_Tutorial1_Solution_1c0165d7.py
@@ -1,13 +1,0 @@
-
-"""
-Discussion:
-
-For the function we have chosen to model the F-I curve (eq 2),
-- a determines the slope (gain) of the rising phase of the F-I curve
-- theta determines the input at which the function F(x) reaches its mid-value (0.5).
-That is, theta shifts the F-I curve along the horizontal axis.
-
-For our neurons we are using in this tutorial:
-- a controls the gain of the neuron population
-- theta controls the threshold at which the neuron population starts to respond
-""";
