+++ tutorials/W2D2_LinearSystems/solutions/W2D2_Tutorial2_Solution_30701e58.py
@@ -1,7 +1,0 @@
-
-# hint: see np.diff()
-inter_switch_intervals = np.diff(switch_times)
-
-# plot inter-switch intervals
-with plt.xkcd():
-  plot_interswitch_interval_histogram(inter_switch_intervals)
