+++ tutorials/W2D4_DynamicNetworks/solutions/W2D4_Tutorial2_Solution_d4eb0391.py
@@ -1,12 +1,0 @@
-
-"""
-Discussion:
-
-- For low values of wEE there is only one fixed point and it is stable so initial
-conditions do not matter and the system always converge to the only fixed point
-
-- For high values of wEE we have three fixed points of which two are stable and
-one is unstable (or saddle). Now it matter where the initial conditions are. If
-the initial conditions are in the attractor region os the high activity fixed
-point then the system will converge to that (the bottom example).
-""";
