+++ tutorials/W1D4_MachineLearning/solutions/W1D4_Tutorial1_Solution_3eaecf0a.py
@@ -1,26 +1,0 @@
-def make_design_matrix(stim, d=25):
-  """Create time-lag design matrix from stimulus intensity vector.
-
-  Args:
-    stim (1D array): Stimulus intensity at each time point.
-    d (number): Number of time lags to use.
-
-  Returns
-    X (2D array): GLM design matrix with shape T, d
-
-  """
-  # Create version of stimulus vector with zeros before onset
-  padded_stim = np.concatenate([np.zeros(d - 1), stim])
-
-  # Construct a matrix where each row has the d frames of
-  # the stimulus proceeding and including timepoint t
-  T = len(stim)  # Total number of timepoints (hint: number of stimulus frames)
-  X = np.zeros((T, d))
-  for t in range(T):
-      X[t] = padded_stim[t:t + d]
-
-  return X
-
-with plt.xkcd():
-  X = make_design_matrix(stim)
-  plot_glm_matrices(X, spikes, nt=50)

[…]
