+++ tutorials/W0D1_PythonWorkshop1/solutions/W0D1_Tutorial1_Solution_5ec18248.py
@@ -1,0 +1,6 @@
+
+# 1. When a is negative, the direction of ax is the exact opposite of x
+# 2. When a is a positive fraction, ax is the same direction but shorter than x
+# 3. Yes you can get z to point anywhere in 2D space by changing a and b around
+# 4. No, if x and y pointed in the same direction, you could not get anywhere in
+#    2D space. We will delve into this in the next section!
