+++ tutorials/W0D3_LinearAlgebra/solutions/W0D3_Tutorial1_Solution_1fbba2ba.py
@@ -1,0 +1,34 @@
+def normalize_vector(input_vector):
+  """ Normalizes a vector to length 1
+
+  Inputs:
+    input_vector (ndarray): array of shape (n_dim, ) where n_dim is the
+        dimensionality of the vector
+
+  Outputs:
+    (ndarray): normalized vector of length 1
+  """
+
+  #################################################################################
+  ## TODO for students: complete normalize_vector function ##
+  # Fill out function and remove
+  # raise NotImplementedError("Student exercise: complete normalize_vector function")
+  #################################################################################
+
+  # Compute vector length (use np.linalg.norm)
+  vector_length = np.linalg.norm(input_vector)
+
+  # Create normalized vector
+  normalized_vector = input_vector / vector_length
+
+  return normalized_vector
+
+# Create vector v from above (use np.array)
+v = np.array([4, 1])
+
+# Call the function
+v_unit = normalize_vector(v)
+
+# Visualize the vectors
+with plt.xkcd():
+  visualize_vectors(v, v_unit)
