+++ tutorials/W0D3_LinearAlgebra/solutions/W0D3_Tutorial1_Solution_ff7c6b75.py
@@ -1,0 +1,10 @@
+
+# 1) They are linearly dependent as one can be formed as a linear combination of
+#     the others (a + b = c). You could also have known this because you have four
+#.    3D vectors. You don't need 4 vectors to get anywhere in 3D space!
+#.
+# 2) The span of a, b, c, and d is all of 3D space (R^3).
+
+# 3) The span of a and b is a 1D line through 3D space (note this is not R^1)
+
+# 4) The span of a and b is a 2d plane through 3D space (note this is not R^2)
