+++ tutorials/W0D3_LinearAlgebra/solutions/W0D3_Tutorial2_Solution_8d909295.py
@@ -1,0 +1,5 @@
+
+A = np.array([[1, 0], [0, 0]])
+
+# Uncomment to visualize transformation
+plot_linear_transformation(A)
