+++ tutorials/youtube_playlists.txt
@@ -1,7 +1,8 @@
 W0D1 https://www.youtube.com/playlist?list=PLkBQOLLbi18M8aIPAL3n2q7D_AbbuPPtM
 W0D2 https://www.youtube.com/playlist?list=PLkBQOLLbi18Oi40BY9HyzVlJFHxhph6Gy
+W0D3 https://www.youtube.com/playlist?list=PLkBQOLLbi18Oz3lb0fo06k0cAWk54rsoE
 W1D1 https://www.youtube.com/playlist?list=PLkBQOLLbi18ObAiSOZ42YBwOQIKNvspeI
-W1D2 https://www.youtube.com/playlist?list=PLkBQOLLbi18Nc7rjBNO99bZQyuTY0TAcE
+W1D2 https://www.youtube.com/playlist?list=PLkBQOLLbi18PxZse6mPHnxviOnp5bMs-g
 W1D3 https://www.youtube.com/playlist?list=PLkBQOLLbi18MqTMnc_mMiuawnczEa8xm9
 W1D4 https://www.youtube.com/playlist?list=PLkBQOLLbi18NQvqJOvyUxeaOwuEICnWB6
 W1D5 https://www.youtube.com/playlist?list=PLkBQOLLbi18N0E0Bx-ZueacF2oRr5eEpU
