+++ tutorials/W2D1_DeepLearning/solutions/W2D1_Tutorial1_Solution_1e3c4101.py
@@ -1,5 +1,6 @@
 
 class DeepNetReLU(nn.Module):
+  """ network with a single hidden layer h with a RELU """
 
   def __init__(self, n_inputs, n_hidden):
     super().__init__()  # needed to invoke the properties of the parent class nn.Module
@@ -21,8 +22,8 @@ torch.manual_seed(1)
 # Get neural responses (r) to and orientation (ori) to one stimulus in dataset
 r, ori = get_data(1, resp_train, stimuli_train)
 
-# Initialize deep network with M=20 hidden units and uncomment lines below
-net = DeepNetReLU(n_neurons, 20)
+# Initialize deep network with M=10 hidden units and uncomment lines below
+net = DeepNetReLU(n_neurons, 10)
 
 # Decode orientation from these neural responses using initialized network
 # net(r) is equivalent to net.forward(r)
