+++ tutorials/W3D1_BayesianDecisions/solutions/W3D1_Tutorial1_Solution_459cbf35.py
@@ -1,0 +1,12 @@
+
+# 1)  With equal probabilities, the expected utility is higher on the left side,
+#    since that is the side without submarines, so you would choose to fish there.
+
+# 2)  If the probability that the fish is on the right side is high, you would
+#     choose to fish there. The high probability of fish being on the right far outweights
+#    the slightly higher utilities from fishing on the left (as you are unlikely to gain these)
+
+# 3)  If the probability that the fish is on the right side is just slightly higher
+#.    than on the left, you would choose the left side as the expected utility is still
+#.    higher on the left. Note that in this situation, you are not simply choosing the
+#.    side with the higher probability - the utility really matters here for the decision
