+++ tutorials/W3D1_BayesianDecisions/solutions/W3D1_Tutorial1_Solution_791e8306.py
@@ -1,0 +1,39 @@
+
+def compute_posterior(likelihood, prior):
+  """ Use Bayes' Rule to compute posterior from likelihood and prior
+
+  Args:
+    likelihood (ndarray): i x j array with likelihood probabilities where i is
+                    number of state options, j is number of measurement options
+    prior (ndarray): i x 1 array with prior probability of each state
+
+  Returns:
+    ndarray: i x j array with posterior probabilities where i is
+            number of state options, j is number of measurement options
+
+  """
+
+  # Compute unnormalized posterior (likelihood times prior)
+  posterior = likelihood * prior # first row is s = left, second row is s = right
+
+  # Compute p(m) (our formula is equivalent to summing each column in the posterior)
+  p_m = np.sum(posterior, axis = 0)
+
+  # Normalize posterior (divide elements by p_m)
+  posterior /= p_m
+
+  return posterior
+
+
+# Create prior
+prior = np.array([0.3, 0.7]).reshape((2, 1)) # first row is s = left, second row is s = right
+
+# Create likelihood
+likelihood = np.array([[0.5, 0.5], [0.1, 0.9]]) # first row is s = left, second row is s = right
+
+# Compute posterior
+posterior = compute_posterior(likelihood, prior)
+
+# Plot all
+with plt.xkcd():
+  plot_prior_likelihood_posterior(prior, likelihood, posterior)
