+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_052f8ff6.py
@@ -1,0 +1,7 @@
+
+""" The limits of the histogram at 0 and 10, as these are the minimum and maximum
+number of trials that a rat can turn left out of 10 trials. The shape seems
+symmetric and is centered around 5. Since an average/mean around 5 left turns out
+of 10 trials implies 5 right turns by the rat on average as well, it is very likely
+the rat chose left and right turns in the maze with equal probability.
+"""
