### 2021-06-01 64df5f3
Add execute flag
For compat with https://github.com/NeuromatchAcademy/nmaci/pull/1
