+++ .github/workflows/notebook-pr.yaml
@@ -70,7 +70,7 @@ jobs:
         run: |
           branch=${{ github.event.pull_request.head.ref }}
           nbs=`python ci/select_notebooks.py ${{ steps.changes.outputs.files }}`
-          python ci/process_notebooks.py $nbs
+          python ci/process_notebooks.py $nbs --execute
           python ci/verify_exercises.py $nbs --c "$COMMIT_MESSAGE"
           python ci/make_pr_comment.py $nbs --branch $branch --o comment.txt
 
