+++ tutorials/W3D1_BayesianDecisions/solutions/W3D1_Tutorial2_Solution_115e84fc.py
@@ -1,0 +1,11 @@
+
+#. 1. Increasing u moves the distribution to the right along the x-axis. The center
+#.    of the distribution equals u - which makes sense as this is the mean! Increasing
+#.    the standard deviation makes the distribution wider.
+
+#.  2. Yes, you can! For example, keep the standard deviation the same and move the mean
+#.     from -2 to 2. At both of these, the probability at 0 is the same because the distribution
+#.     is symmetrical.
+
+#.  3. There are an infinite number of Gaussians (combinations of mean & standard deviation)
+#.     that could produce the same probability at 0
