+++ tutorials/W3D1_BayesianDecisions/solutions/W3D1_Tutorial2_Solution_043fbb5f.py
@@ -1,0 +1,14 @@
+
+#. 1) As Astrocat's mean get closer to the mean of the gain (or loss), the EU become dominated
+#.    by only the gain or loss.
+
+#. 2) Only if the mean and variances of both the gain and loss regions are exactly the same.
+# .   (Set one of the variances 0.01 more than the other to see this.)
+
+#. 3) If the variances of the gain and loss function are small enough relative to the position
+#.    of Astrocat, there will be a 'neutral' region. As the variances increase, this will go away.
+
+#. 4) As the uncertainty of Astrocat's location increases (relative to the gain and loss variances).
+#.    there will be a continuous increase in utility from the peak of the loss region to the peak of the
+#.    gain region. Also, this will depend on the mean of Astrocat's distribution! The larger the variance
+#.    the more sensitive the expected utility is to both the gains and losses!
