+++ tutorials/W0D4_Calculus/solutions/W0D4_Tutorial1_Solution_fd942e45.py
@@ -1,0 +1,13 @@
+
+"""
+1. The smallest value of dt results in the best numerical integration (most similar
+   to the analytical solution)
+
+2. We have to compute the area of more rectangles.
+
+3. Here we used the forward Riemann sum, and it results in an underestimate of the
+   integral for positive values of $T$ and underestimate for negativeve values of
+   $T$. We could also use the backward Riemann sum, and that will give an overestimate
+   of the integral for positive values of $T$.
+
+"""
