+++ tutorials/W0D4_Calculus/solutions/W0D4_Tutorial2_Solution_fe6aea6f.py
@@ -1,0 +1,6 @@
+
+"""
+  1. For $V>-75$ mV, the derivative is negative.
+  2. For $V<-75$ mV, the derivative is positive.
+  3. For $V=-75$ mV, the derivative is equal to $0$ is and a stable point.
+"""
