+++ tutorials/W0D4_Calculus/solutions/W0D4_Tutorial2_Solution_9953ed40.py
@@ -1,0 +1,10 @@
+p = np.arange(0, 100, 0.1)
+
+with plt.xkcd():
+
+  dpdt = 0.3*p
+  fig = plt.figure(figsize=(6, 4))
+  plt.plot(p, dpdt)
+  plt.xlabel(r'Population $p(t)$ (millions)')
+  plt.ylabel(r'$\frac{d}{dt}p(t)=\alpha p(t)$')
+  plt.show()
