+++ tutorials/W0D4_Calculus/solutions/W0D4_Tutorial1_Solution_a0e42694.py
@@ -1,0 +1,17 @@
+
+"""
+
+We can set t equal to the exponent:
+t = a^4 + 1
+r(t) = e^t
+
+
+Then:
+ dt/da = 4a^3
+ dr/dt = e^t
+
+Now we can use the chain rule:
+dr/da = dr/dt * dt/da
+      = e^t(4a^3)
+      = 4a^3e^{a^4 + 1}
+"""
