### 2021-06-02 #110 @emckiernan
Hi, @eglerean, you previously signed up to review for JOSE. Would you please consider contributing a review for this submission? The title is "Neuromatch Academy: a 3-week, online summer school in computational neuroscience". Please let me know if you're available and willing. Thanks!
