+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_90422623.py
@@ -1,0 +1,10 @@
+
+# Set random seed
+np.random.seed(0)
+
+# Draw 5 samples from a Poisson distribution with lambda = 4
+sampled_spike_counts = np.random.poisson(4, 5)
+
+# Print the counts
+print("The samples drawn from the Poisson distribution are " +
+          str(sampled_spike_counts))
