+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_6864c270.py
@@ -1,10 +1,0 @@
-
-""" Have you noticed that the distribution starts to look more symmetric with
-more samples? This is a result of the Central Limit Theorem, where a large number
-of sampled values makes samples from the Poisson distribution look similarly
-distributed to the symmetric Gaussian (Normal) distribution.
-
-Increasing lambda also makes the spread of samples from a poisson distribution more
-symmetric and able to be approximated by the normal distribution. This approximation
-gets better and better with larger sample sizes, and larger values of lambda.
-"""
