+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_606c3f63.py
@@ -1,3 +1,0 @@
-
-""" To be added!
-"""
