+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_6e912c89.py
@@ -1,0 +1,8 @@
+
+"""
+1) A larger step size could mean that the rat is moving faster, or that we sample
+   the rats location less often.
+
+2) The rat tends to visit more of the arena with a large step size.
+
+"""
