+++ tutorials/W3D2_HiddenDynamics/solutions/W3D2_Tutorial1_Solution_a956a168.py
@@ -1,0 +1,10 @@
+
+"""
+
+1) Low levels of noise results in higher accuracies generally, especially
+   at early stop times.
+
+2) High levels of noise results in lower accuracies generally.
+
+
+"""
