+++ tutorials/W3D2_HiddenDynamics/solutions/W3D2_Tutorial2_Solution_c36f28a6.py
@@ -4,7 +4,7 @@ switch_prob = 0.1
 log10_noise_level = -1
 
 # Build model
-model = create_model(switch_prob=switch_prob,
+model = create_HMM(switch_prob=switch_prob,
                      noise_level=10.**log10_noise_level,
                      startprob=[0.5, 0.5])
 observations, states = model.sample(nstep)
