+++ tutorials/youtube_playlists.txt
@@ -1,6 +1,7 @@
 W0D1 https://www.youtube.com/playlist?list=PLkBQOLLbi18M8aIPAL3n2q7D_AbbuPPtM
 W0D2 https://www.youtube.com/playlist?list=PLkBQOLLbi18Oi40BY9HyzVlJFHxhph6Gy
 W0D3 https://www.youtube.com/playlist?list=PLkBQOLLbi18Oz3lb0fo06k0cAWk54rsoE
+W0D5 https://www.youtube.com/playlist?list=PLkBQOLLbi18MDDH09-kvRoU8XANuCJvGg
 W1D1 https://www.youtube.com/playlist?list=PLkBQOLLbi18ObAiSOZ42YBwOQIKNvspeI
 W1D2 https://www.youtube.com/playlist?list=PLkBQOLLbi18PxZse6mPHnxviOnp5bMs-g
 W1D3 https://www.youtube.com/playlist?list=PLkBQOLLbi18MqTMnc_mMiuawnczEa8xm9
