### 2021-06-03 c03e241
W0D5 updates (#569)
* Created using Colaboratory

* Created using Colaboratory

* Created using Colaboratory

* Delete W0D5_Tutorial1.ipynb

* Tutorial 1 post production editing

* Tutorial 2 post production editing

* Tutorial 1 Post Production Editing Update

* Tutorial 1 Post Production Editing Update

* Tutorial 2 Post Production Editing Update

* Tutorial 2 Post Production Editing Update

* Tutorial 2 Post Production Editing Update

* Process tutorial notebooks

* Final small edits to pyflakes errors

* Final small edits to pyflakes errors

* Process tutorial notebooks

* Edit exercises


[…]
