+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial2_Solution_daa12602.py
@@ -1,0 +1,10 @@
+
+""" You will learn more about "Bayesian brains" and the theory surrounding
+these ideas once the course begins. Here is a brief explanation: it may
+be ideal for human brains to implement Bayesian inference by integrating "prior"
+information the brain has about the world (memories, prior knowledge, etc.) with
+new evidence that updates its "beliefs"/prior. This process seems to parallel
+the brain's method of learning about its environment, making it a compelling
+theory for many neuroscience researchers. The next exercise examines a possible
+real world model for Bayesian inference: sound localization.
+"""
