+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial2_Solution_4cfc70ca.py
@@ -1,0 +1,55 @@
+
+def classic_vs_bayesian_normal(mu, sigma, num_points, prior):
+  """ Compute both classical and Bayesian inference processes over the range of
+  data sample sizes (num_points) for a normal distribution with parameters
+  mu,sigma for comparison.
+
+  Args:
+    mu (scalar): the mean parameter of the normal distribution
+    sigma (scalar): the standard deviation parameter of the normal distribution
+    num_points (int): max number of points to use for inference
+    prior (ndarray): prior data points for Bayesian inference
+
+  Returns:
+    mean_classic (ndarray): estimate mean parameter via classic inference
+    var_classic (ndarray): estimate variance parameter via classic inference
+    mean_bayes (ndarray): estimate mean parameter via Bayesian inference
+    var_bayes (ndarray): estimate variance parameter via Bayesian inference
+  """
+
+  # Initialize the classical and Bayesian inference arrays that will estimate
+  # the normal parameters given a certain number of randomly sampled data points
+  mean_classic = np.zeros(num_points)
+  var_classic = np.zeros(num_points)
+
+  mean_bayes = np.zeros(num_points)
+  var_bayes = np.zeros(num_points)
+
+  for nData in range(num_points):
+    # Randomly sample nData + 1 number of points
+    x = generate_normal_samples(mu, sigma, nData + 1)
+    # Compute the mean of those points and set the corresponding array entry to this value
+    mean_classic[nData] = np.mean(x)
+    # Compute the variance of those points and set the corresponding array entry to this value
+    var_classic[nData] = np.var(x)
+
+    # Bayesian inference with the given prior is performed below for you
+    xsupp = np.hstack((x, prior))
+    mean_bayes[nData] = np.mean(xsupp)
+    var_bayes[nData] = np.var(xsupp)
+
+  return mean_classic, var_classic, mean_bayes, var_bayes
+
+# Set normal distribution parameters, mu and sigma
+mu = 5
+sigma = 1
+
+# Set the prior to be two new data points, 5 and 7, and print the mean and variance
+prior = np.array((5, 7))
+print("The mean of the data comprising the prior is: " + str(np.mean(prior)))
+print("The variance of the data comprising the prior is: " + str(np.var(prior)))
+
+# Uncomment once the function above is completed
+mean_classic, var_classic, mean_bayes, var_bayes = classic_vs_bayesian_normal(mu, sigma, 30, prior)
+with plt.xkcd():
+  plot_classical_vs_bayesian_normal(30, mean_classic, var_classic, mean_bayes, var_bayes)
