+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_06a79c9b.py
@@ -1,0 +1,14 @@
+
+"""
+1) The limits of the histogram at 0 and 10, as these are the minimum and maximum
+number of trials that a rat can turn left out of 10 trials.
+
+2) The shape seems symmetric and is centered around 5.
+
+3) An average/mean around 5 left turns out of 10 trials indicates that the rat
+chose left and right turns in the maze with equal probability (that p = 0.5)
+
+4) With p = 0.8, the center of the histogram would be at x = 8 and it would not be
+   as symmetrical (since it would be cut off at max 10). You can go into the code above
+   and run it with p = 0.8 to see this.
+"""
