+++ tutorials/W0D5_Statistics/solutions/W0D5_Tutorial1_Solution_f415a4a6.py
@@ -1,0 +1,28 @@
+def my_gaussian(x_points, mu, sigma):
+  """ Returns normalized Gaussian estimated at points `x_points`, with
+  parameters: mean `mu` and standard deviation `sigma`
+
+  Args:
+      x_points (ndarray of floats): points at which the gaussian is evaluated
+      mu (scalar): mean of the Gaussian
+      sigma (scalar): standard deviation of the gaussian
+
+  Returns:
+      (numpy array of floats) : normalized Gaussian evaluated at `x`
+  """
+
+  px = 1/(2*np.pi*sigma**2)**1/2 *np.exp(-(x_points-mu)**2/(2*sigma**2))
+
+  # as we are doing numerical integration we may have to remember to normalise
+  # taking into account the stepsize (0.1)
+  px = px/(0.1*sum(px))
+  return px
+
+x = np.arange(-8, 9, 0.1)
+
+# Generate Gaussian
+px = my_gaussian(x, -1, 1)
+
+# Visualize
+with plt.xkcd():
+  my_plot_single(x, px)

[…]
