### 2021-06-03 #110 @djmannion
> Hi, @djmannion, you previously signed up to review for JOSE. Would you please consider contributing a review for this submission? The title is "Neuromatch Academy: a 3-week, online summer school in computational neuroscience". Please let me know if you're available and willing. Thanks!

Hi @emckiernan. Thank you for the invitation, but I am unfortunately unavailable for reviewing manuscripts at the moment - apologies.
