+++ tutorials/W3D2_HiddenDynamics/solutions/W3D2_Tutorial3_Solution_8bda671a.py
@@ -1,0 +1,12 @@
+
+"""
+
+1) sigma_m is how much noise there is in the measurement observation model
+   (the reading of the color location).
+   sigma_p was the noise in the actual movement every step.
+
+2) Accuracy will be worse with higher noise
+
+3) If sigma_m is 0, the measured states will be fully accurate.
+
+"""
