### 2021-06-03 0fc5e1a
W3D2 Updates (#577)
* Update to 2021 content

* Refresh kernel

* Sync exercise comments

* Process tutorial notebooks

* Update to 2021 content

* Fix solution

* Fix wrong name for function

* Sync solutions

* Process tutorial notebooks

* Update to 2021 content

* Update to 2021 content

* Refresh kernel

* Process tutorial notebooks

* Update to 2021 content

* Sync solution comments

* Process tutorial notebooks

Co-authored-by: GitHub Action <action@github.com>
