+++ tutorials/README.md
@@ -52,6 +52,7 @@
 
 ## W0D5 - Statistics
 
+[YouTube Playlist](https://www.youtube.com/playlist?list=PLkBQOLLbi18MDDH09-kvRoU8XANuCJvGg)
 |   | Run | View |
 | - | --- | ---- |
 | Tutorial 1 | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/tutorials/W0D5_Statistics/student/W0D5_Tutorial1.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/tutorials/W0D5_Statistics/student/W0D5_Tutorial1.ipynb?flush_cache=true) |
