### 2021-06-03 #110 @mnarayan
I am unfortunately not available for reviewing this month.

On Thu, Jun 3, 2021 at 10:02 AM Erin C. McKiernan ***@***.***>
wrote:

> Hi, @mnarayan <https://github.com/mnarayan>, you previously signed up to
> review for JOSE. Would you please consider contributing a review for this
> submission? The title is "Neuromatch Academy: a 3-week, online summer
> school in computational neuroscience". Please let me know if you're
> available and willing. Thanks!
>
