+++ tutorials/materials.yml
@@ -1,10 +1,5 @@
-- day: W0D1
-  name: Python Workshop 1
-  tutorials: 1
-- day: W0D2
-  name: Python Workshop 2
-  tutorials: 1
 - day: W1D1
+  category: Intro to Modeling
   intro: https://www.youtube.com/watch?v=KxldhMR5PxA
   name: Model Types
   outro: https://www.youtube.com/watch?v=KZQXfQL1SH4
@@ -26,10 +21,11 @@
     title: TA
   tutorials: 3
 - day: W1D2
-  intro: https://www.youtube.com/watch?v=8pz_NH5_Zy4
+  category: Intro to Modeling
+  intro: https://www.youtube.com/watch?v=65HhPd0kG2k
   name: Modeling Practice
-  outro: https://www.youtube.com/watch?v=Il8zOmCMFAA
-  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18Nc7rjBNO99bZQyuTY0TAcE
+  outro: https://www.youtube.com/watch?v=JSTjK6PrOgY
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PxZse6mPHnxviOnp5bMs-g
   qa:
   - https://www.youtube.com/watch?v=j8xsEOg8F_Q
   - https://www.youtube.com/watch?v=VAM-suHZPSM
@@ -47,6 +43,7 @@
     title: TA
   tutorials: 2
 - day: W1D3
+  category: Machine Learning
   intro: https://www.youtube.com/watch?v=yTQNqYCvoZ0
   name: Model Fitting
   outro: https://www.youtube.com/watch?v=cUAUL3efqpY
@@ -66,6 +63,7 @@
     title: Reading
   tutorials: 6
 - day: W1D4
+  category: Machine Learning
   intro: https://www.youtube.com/watch?v=m1w7oywzwpA
   name: Machine Learning
   outro: https://www.youtube.com/watch?v=NXVG9ORBYXQ
@@ -85,6 +83,7 @@
     title: Reading
   tutorials: 2
 - day: W1D5
+  category: Machine Learning
   intro: https://www.youtube.com/watch?v=zeBFyRaoVnQ
   name: Dimensionality Reduction
   outro: https://www.youtube.com/watch?v=Vaut2BJlv10
@@ -103,26 +102,8 @@
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/sjqg5/?direct%26mode=render%26action=download%26mode=render
     title: Reading
   tutorials: 4
-- day: W2D1
-  intro: https://www.youtube.com/watch?v=uW8dbFbdOjo
-  name: Bayesian Statistics
-  outro: https://www.youtube.com/watch?v=njSGa9__fF0
-  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18NmYKYd6AjspUPnlys9R3R0
-  qa:
-  - https://www.youtube.com/watch?v=CVXe6fJ9xdI
-  - https://www.youtube.com/watch?v=b8G2cfMrHjY
-  - https://www.youtube.com/watch?v=Un9yDx5vzIs
-  slides:
-  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/xam4w/?direct%26mode=render%26action=download%26mode=render
-    title: Intro
-  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/sqnd5/?direct%26mode=render%26action=download%26mode=render
-    title: Tutorials
-  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/cvbu7/?direct%26mode=render%26action=download%26mode=render
-    title: Outro
[…]
