### 2021-06-03 #118 @whedon
[REVIEW]: Neuromatch Academy: a 3-week, online summer school in computational neuroscience
**Submitting author:** @thartbm (<a href="http://orcid.org/0000-0002-0045-3769">Bernard Marius &#39;t Hart</a>)
**Repository:** <a href="https://github.com/NeuromatchAcademy/course-content/" target ="_blank">https://github.com/NeuromatchAcademy/course-content/</a>
**Version:** NMA-CN-2021
**Archive:** <a href="https://doi.org/10.17605/OSF.IO/CFN4U" target="_blank">10.17605/OSF.IO/CFN4U</a>

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".

- [x] **Version:** NMA-CN-2021

- [ ] **Version:** NMA-CN-2021
