### 2021-06-03 #118 @labarba
The repository given by the authors is  https://github.com/NeuromatchAcademy/course-content/tree/jose_paper
... but we need https://github.com/NeuromatchAcademy/course-content/
... to then proceed to compile from branch `jose_paper`.

Looking at the [Pre-Review](https://github.com/openjournals/jose-reviews/issues/110#event-4839203177) issue, where the repo URL is correct, my best guess is that I edited the repo by hand on the top whedon post. I guess the author-entered URL remains attached to the submission on the backend.
