### 2021-06-03 7103a8e
manual build of book
