### 2021-06-03 8779b25
update for first real draft of book
