+++ tutorials/_config.yml
@@ -8,14 +8,14 @@
 title                       : Neuromatch Computational Neuroscience  # The title of the book. Will be placed in the left navbar.
 author                      : Neuromatch  # The author of the book
 copyright                   : "2021"  # Copyright year to be placed in the footer
-#logo                        : nma_cn_logo.png  # A path to the book logo
+logo                        : nma-logo-square-4xp.jpg  # A path to the book logo
 only_build_toc_files        : true
 
 
 #######################################################################################
 # Execution settings
 execute:
-  execute_notebooks         : cache # Whether to execute notebooks at build time. Must be one of ("auto", "force", "cache", "off")
+  execute_notebooks         : off # Whether to execute notebooks at build time. Must be one of ("auto", "force", "cache", "off")
   cache                     : ""    # A path to the jupyter cache that will be used to store execution artifacts. Defaults to `_build/.jupyter_cache/`
   exclude_patterns          : []    # A list of patterns to *skip* in execution (e.g. a notebook that takes a really long time)
   timeout                   : 30    # The maximum time (in seconds) each notebook cell is allowed to run.
