### 2021-06-03 0aa7330
update for first real draft of book
