+++ ci/generate_book.py
@@ -5,9 +5,11 @@ def main():
     with open('tutorials/materials.yml') as fh:
         materials = yaml.load(fh, Loader=yaml.FullLoader)
     
-    toc = {'Intro to Modeling': {'part': 'Intro to Modeling', 'chapters': []},
+    toc = {'Pre-reqs Refresher': {'part': 'Pre-reqs Refresher', 'chapters': []},
+           'Intro to Modeling': {'part': 'Intro to Modeling', 'chapters': []},
            'Machine Learning': {'part': 'Machine Learning', 'chapters': []},
-           'Dynamical Systems': {'part': 'Dynamical Systems', 'chapters': []}}
+           'Dynamical Systems': {'part': 'Dynamical Systems', 'chapters': []},
+           'Stochastic Processes': {'part': 'Stochastic Processes', 'chapters': []}}
     
     for m in materials:
         directory = f"{m['day']}_{''.join(m['name'].split())}"
@@ -16,11 +18,16 @@ def main():
                    'sections': []}
         print(m['day'])
         part = m['category']
-        
+
+        # Make intro video page
+
+        # Add tutorials
         for i in range(m['tutorials']):
             directory = f"{m['day']}_{''.join(m['name'].split())}"
             notebook = f"{m['day']}_Tutorial{i+1}.ipynb"
             chapter['sections'].append({'file': f"{directory}/student/{notebook}"})
+
+        # Make outro video page
 
         toc[part]['chapters'].append(chapter)
 
