+++ tutorials/materials.yml
@@ -1,3 +1,23 @@
+- day: W0D1
+  category: Pre-reqs Refresher
+  name: Python Workshop 1
+  tutorials: 1
+- day: W0D2
+  category: Pre-reqs Refresher
+  name: Python Workshop 2
+  tutorials: 1
+- day: W0D3
+  category: Pre-reqs Refresher
+  name: Linear Algebra
+  tutorials: 3
+- day: W0D4
+  category: Pre-reqs Refresher
+  name: Calculus
+  tutorials: 3
+- day: W0D5
+  category: Pre-reqs Refresher
+  name: Statistics
+  tutorials: 2
 - day: W1D1
   category: Intro to Modeling
   intro: https://www.youtube.com/watch?v=KxldhMR5PxA
@@ -102,6 +122,26 @@
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/sjqg5/?direct%26mode=render%26action=download%26mode=render
     title: Reading
   tutorials: 4
+- day: W2D1
+  category: Machine Learning
+  intro: https://www.youtube.com/watch?v=IZvcy0Myb3M
+  name: Deep Learning
+  outro: https://www.youtube.com/watch?v=nWlgIclpyt4
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PZ2uw0p7G4EkzjzqP8l0Eg
+  qa:
+  - https://www.youtube.com/watch?v=c7DftQoCcf8
+  - https://www.youtube.com/watch?v=eXhPvvAXVmc
+  - https://www.youtube.com/watch?v=lYl59QduHJQ
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/jw829/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/fsqhr/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/z5g93/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/w6dfq/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  tutorials: 3
 - day: W2D2
   category: Dynamical Systems
   intro: https://www.youtube.com/watch?v=8NsLc0KoeYg
@@ -120,5 +160,145 @@
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/dx6je/?direct%26mode=render%26action=download%26mode=render
     title: Outro
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/dcrnj/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  tutorials: 4
+- day: W2D3
+  category: Dynamical Systems
+  intro: https://www.youtube.com/watch?v=MAOOPv3whZ0
+  name: Real Neurons
+  outro: https://www.youtube.com/watch?v=pMxb9vA-_4U
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18MCEdPJQ7gdnqP-Z0Tkmcjy
+  qa:
+  - https://www.youtube.com/watch?v=KYwer_JekG8
+  - https://www.youtube.com/watch?v=Qxmylyyxyww
+  - https://www.youtube.com/watch?v=fJQXi-kGqQk
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/gyfr2/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
[…]
