### 2021-06-04 #118 @arfon
> Problem with affiliations for Xaq Pitkow, perhaps the affiliations index need quoting?

This is your problem. Xaq Pitkow's affiliations need putting in quotes.
