### 2021-06-04 #118 @thartbm
OK, happy to change. Should it be like this:

` - name: Xaq Pitkow
   orcid: 0000-0001-6376-329X
   affiliation: "37,43"`

Like this:

` - name: Xaq Pitkow
   orcid: 0000-0001-6376-329X
   affiliation: "37","43"`

Or like this:

` - name: "Rice University"
   index: 37
...
 - name: "Baylor College of Medicine"
   index: 43`

Or some other way?
