### 2021-06-04 #118 @arfon
As [per our docs](https://openjournals.readthedocs.io/en/jose/submitting.html#example-paper-and-bibliography):

```
 - name: Xaq Pitkow
   orcid: 0000-0001-6376-329X
   affiliation: "37,43"
```
