### 2021-06-07 953bbe4
Add environment variable for https://github.com/NeuromatchAcademy/nmaci/pull/6
