+++ .github/workflows/notebook-pr.yaml
@@ -6,6 +6,7 @@ on:
 
 env:
   NB_KERNEL: python
+  NMA_MAIN_BRANCH: master
 
 jobs:
 
