### 2021-06-07 f7dfc06
added quotations around Xaq Pitkow's affiliations
documentation:
https://openjournals.readthedocs.io/en/jose/submitting.html#example-paper-and-bibliography
