### 2021-06-11 548d0d5
fix rendering of amsmath latex environments (#582)
