### 2021-06-11 0d40e39
Move book creation to main folder so we can include projects
