+++ _config.yml
@@ -8,7 +8,7 @@
 title                       : Neuromatch Computational Neuroscience  # The title of the book. Will be placed in the left navbar.
 author                      : Neuromatch  # The author of the book
 copyright                   : "2021"  # Copyright year to be placed in the footer
-logo                        : nma-logo-square-4xp.jpg  # A path to the book logo
+logo                        : tutorials/nma-logo-square-4xp.jpg  # A path to the book logo
 only_build_toc_files        : true
 
 
@@ -41,5 +41,5 @@ launch_buttons:
 
 repository:
   url                       : https://github.com/NeuromatchAcademy/course_content  # The URL to your book's repository
-  path_to_book              : book  # A path to your book's folder, relative to the repository root.
+  path_to_book              : ./  # A path to your book's folder, relative to the repository root.
   branch                    : master  # Which branch of the repository should be used when creating links
