### 2021-06-11 9f1f8eb
Move book creation to main folder so we can include projects
