+++ README.md
@@ -6,8 +6,6 @@
 
 **Objectives**: Introduce traditional and emerging computational neuroscience tools, their complementarity, and what they can tell us about the brain. A main focus is on modeling choices, model creation, model evaluation and understanding how they relate to biological questions.
 
-**Tutorial microstructure**: ~5min talk, ~15min hands-on (repeated)
-
 **Day structure**: Opening keynote, 3h15min lecture/live Q&A/tutorial modules, 40min group discussion, 55min interpretation lecture (Outro) + live Q&As (what did we learn today, what does it mean, underlying philosophy). There will also be many networking activities (professional development sessions, yoga, social hangouts etc)! Details below.
 
 **Note for visitors from China**: This repository contains many links to YouTube and Google Colab. We have a version of the repository with those same videos posted on bilibili, and the Google Colab links replaced with links to Aliyun. Please visit the [China Accessible Neuromatch Course-Content](https://github.com/erlichlab/course-content/)
@@ -25,453 +23,35 @@ Group projects are offered for the *interactive track* only and will be running 
 # Course outline
 
 * [Week 0](#week-0)
-    * [Python for NMA Part 1](#python-for-nma-workshop)
-    * [Python for NMA Part 2](#python-for-nma-workshop) 
+    * [Python Workshop Part 1]
+    * [Python Workshop Part 2]
     * [Linear Algebra]() 
     * [Calculus]() 
     * [Probability & Statistics]() 
 * [Week 1](#week-1)
-    * [Mon, July 13: Model Types](#mon-july-13-model-types)
-    * [Tue, July 14: Modeling Practice](#tue-july-14-modeling-practice)
-    * [Wed, July 15: Model Fitting](#wed-july-15-model-fitting)
-    * [Thu, July 16: Machine Learning: GLM](#thu-july-16-machine-learning)
-    * [Fri, July 17: Dimensionality Reduction](#fri-july-17-dimensionality-reduction)
+    * [Mon, July 5: Model Types]
+    * [Tue, July 6: Modeling Practice]
+    * [Wed, July 7: Model Fitting]
+    * [Thu, July 8: Generalized Linear Models]
+    * [Fri, July 9: Dimensionality Reduction]
     
- * [Professional Development](#weekend-1-july-18--19)
+ * [Professional Development]
     
 * [Week 2](#week-2)
-    * [Mon, July 20: Bayesian Statistics](#mon-july-20-bayesian-statistics)
-    * [Tue, July 21: Linear Systems](#tue-july-21-linear-systems)
-    * [Wed, July 22: Decision Making](#wed-july-22-decision-making)
-    * [Thu, July 23: Optimal Control](#thu-july-23-optimal-control)
-    * [Fri, July 24: Reinforcement Learning](#fri-july-24-reinforcement-learning)
+    * [Mon, July 12: Deep Learning]
+    * [Tue, July 13: Linear Systems]
+    * [Wed, July 14: Biological Neuron Models]
+    * [Thu, July 15: Dynamic Networks]
+    * [Fri, July 16: Project day!]
     
- * [Professional Development](#weekend-2-july-25--26)
+ * [Professional Development]
  
 * [Week 3](#week-3)
-    * [Mon, July 27: Real Neurons](#mon-july-27-real-neurons)
-    * [Tue, July 28: Dynamic Networks](#tue-july-28-dynamic-networks)
-    * [Wed, July 29: Network Causality](#wed-july-29-network-causality)    
-    * [Thu, July 30: Deep Learning 1](#thu-july-30-deep-learning-1)
-    * [Fri, July 31: Deep Learning 2](#fri-july-31-deep-learning-2)
-
-----
-
-## Week 0
-
-### Python for NMA workshop
-
[…]
