### 2021-06-14 ce2cd7c
Update Readme to 2021 schedule, remove detailed day schedules for now (#585)
* Update to 2021 schedule

* Update README.md

* Update README.md

* Update README.md
