+++ ci/resources/intro_outro_template.txt
@@ -1,0 +1,7 @@
+# {{type}} Video
+## Video
+
+<iframe width="420" height="315" src="{{video_source_url}}"></iframe>
+
+## Slides
+<iframe src="{{slide_source_url}}", frameborder="0" width="960" height="569" allowfullscreen="true" mozallowfullscreen="true" webkitallowfullscreen="true"></iframe>
