+++ tutorials/W2D1_DeepLearning/intro_vid.md
@@ -1,0 +1,7 @@
+# Intro Video
+## Video
+
+<iframe width="420" height="315" src="https://www.youtube.com/embed/IZvcy0Myb3M"></iframe>
+
+## Slides
+<iframe src="https://mfr.ca-1.osf.io/render?url=https://osf.io/jw829/?direct%26mode=render%26action=download%26mode=render", frameborder="0" width="960" height="569" allowfullscreen="true" mozallowfullscreen="true" webkitallowfullscreen="true"></iframe>
