+++ projects/README.md
@@ -89,7 +89,7 @@ publication-worthy project! All groups must finish their projects by the end of 
  * [Example project video two](https://www.youtube.com/watch?v=pvfy0_v12wI)
  * PROJECTS ARE DUE PRIOR TO THE INTRO FOR W3D5!!
  * Submit your project (only once per group) using this [project submission form](https://airtable.com/shrBTUnFNkVcHSgl5)
- 
+
  * **NMA final day** — Presentation of projects
 
   * The afternoon of the final day of NMA will include project presentations!
@@ -120,12 +120,12 @@ publication-worthy project! All groups must finish their projects by the end of 
 
 |   | Run | View |
 | - | --- | ---- |
-| Steinmetz decisions | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_steinmetz_decisions.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_steinmetz_decisions.ipynb)|
-| Steinmetz extra | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_steinmetz_extra.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_steinmetz_extra.ipynb)|
-| Stringer spontaneous | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_stringer_spontaneous.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_stringer_spontaneous.ipynb)|
-| Stringer orientations | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_stringer_orientations.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_stringer_orientations.ipynb) |
-| Kay images | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_kay_images.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_kay_images.ipynb) |
-| HCP | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/projects/load_hcp.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/projects/load_hcp.ipynb) |
[…]
