### 2021-06-18 #118 @PaulScotti
> 👋 @PaulScotti, please update us on how your review is going (this is an automated reminder).

Hoping to get to this sometime next week
