### 2021-06-18 105ebbd
move book things into book directory; autopublish workflow
