+++ tutorials/W3D1_BayesianDecisions/intro_vid.md
@@ -1,7 +1,0 @@
-# Intro Video
-## Video
-
-<iframe width="420" height="315" src="https://www.youtube.com/embed/w2GH1-Iqods"></iframe>
-
-## Slides
-<iframe src="https://mfr.ca-1.osf.io/render?url=https://osf.io/xam4w/?direct%26mode=render%26action=download%26mode=render", frameborder="0" width="960" height="569" allowfullscreen="true" mozallowfullscreen="true" webkitallowfullscreen="true"></iframe>
