+++ tutorials/W1D1_ModelTypes/outro_vid.md
@@ -1,7 +1,0 @@
-# Outro Video
-## Video
-
-<iframe width="420" height="315" src="https://www.youtube.com/embed/KZQXfQL1SH4"></iframe>
-
-## Slides
-<iframe src="https://mfr.ca-1.osf.io/render?url=https://osf.io/9hkg2/?direct%26mode=render%26action=download%26mode=render", frameborder="0" width="960" height="569" allowfullscreen="true" mozallowfullscreen="true" webkitallowfullscreen="true"></iframe>
