### 2021-06-19 fb6eb3a
Adding code to open in colab in new tab
