### 2021-06-19 46fb9be
Create jupyterbook  (#583)
Added support for jupyterbook binding of all the tutorials w/ automatic publish to the gh-pages branch on every merge into master
