+++ book/intro.md
@@ -1,0 +1,3 @@
+# Introduction
+
+Welcome to the Neuromatch computational neuroscience course!
