+++ tutorials/materials.yml
@@ -1,0 +1,304 @@
+- day: W0D1
+  category: Pre-reqs Refresher
+  name: Python Workshop 1
+  tutorials: 1
+- day: W0D2
+  category: Pre-reqs Refresher
+  name: Python Workshop 2
+  tutorials: 1
+- day: W0D3
+  category: Pre-reqs Refresher
+  name: Linear Algebra
+  tutorials: 3
+- day: W0D4
+  category: Pre-reqs Refresher
+  name: Calculus
+  tutorials: 3
+- day: W0D5
+  category: Pre-reqs Refresher
+  name: Statistics
+  tutorials: 2
+- day: W1D1
+  category: Intro to Modeling
+  intro: https://www.youtube.com/watch?v=KxldhMR5PxA
+  name: Model Types
+  outro: https://www.youtube.com/watch?v=KZQXfQL1SH4
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18ObAiSOZ42YBwOQIKNvspeI
+  qa:
+  - https://www.youtube.com/watch?v=FwsKmunwzUA
+  - https://www.youtube.com/watch?v=h63bKG0-h_w
+  - https://www.youtube.com/watch?v=8h0X9lxY8Bs
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/rbx2a/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/6dxwe/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/9hkg2/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/2esh5/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/5vj73/?direct%26mode=render%26action=download%26mode=render
+    title: TA
+  tutorials: 3
+- day: W1D2
+  category: Intro to Modeling
+  intro: https://www.youtube.com/watch?v=65HhPd0kG2k
+  name: Modeling Practice
+  outro: https://www.youtube.com/watch?v=JSTjK6PrOgY
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PxZse6mPHnxviOnp5bMs-g
+  qa:
+  - https://www.youtube.com/watch?v=j8xsEOg8F_Q
+  - https://www.youtube.com/watch?v=VAM-suHZPSM
+  - https://www.youtube.com/watch?v=ha72Q_0e3mA
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/kmwus/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/kygfn/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/agrp6/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/2esh5/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/v5emn/?direct%26mode=render%26action=download%26mode=render
+    title: TA
+  tutorials: 2
+- day: W1D3
[…]

[…]
