### 2021-06-19 8b27b5f
Update course book
