+++ tutorials/W3D1_BayesianDecisions/intro_text.md
@@ -1,8 +1,3 @@
-# W3D1 - Bayesian Decisions
+# Bayesian Decisions
 
-Paragraph 1. A very brief paragraph overviewing the day
-
-Paragraph 2. High level overview of how this content fits in the field of comp neuro/course, how it builds on previous days, how it will connect to future days. This can be pretty broad and brief. 
-
-Paragraph 3. A couple examples of what sorts of research questions the content is useful for. Again, this is very basic and introductory so no details necessary. 
-
+Today, we'll be learning about Bayesian Decisions.
