### 2021-06-22 2a09912
Code to hide cells in jupyterbook (#592)
* Adding code for open in colab in new tab

* Adding code to hide colab hidden cells in jupyterbook
