+++ tutorials/W1D4_MachineLearning/further_reading.md
@@ -1,0 +1,92 @@
+# Generalized Linear Models
+
+## Suggested further readings
+
+&nbsp;
+
+McCullagh, P., & Nelder, J. A. (1989). Generalized linear models. Chapman & Hall.
+
+Simoncelli, E. P., Paninski, L., Pillow, J., & Schwartz, O. (2004). Characterization of Neural
+Responses with Stochastic Stimuli. In M. S. Gazzaniga (Ed.), The Cognitive Neurosciences III (p.
+327). The MIT Press.
+
+Paninski, L. (2004). Maximum likelihood estimation of cascade point-process neural encoding
+models. Network: Comput. Neural Syst., 15, 243–262.
+
+Truccolo, W., Eden, U. T., Fellows, M. R., Donoghue, J. P., & Brown, E. N. (2005). A Point
+Process Framework for Relating Neural Spiking Activity to Spiking History, Neural Ensemble, and
+Extrinsic Covariate Effects. Journal of Neurophysiology, 93(2), 1074–1089.
+
+Pillow, J. W., Paninski, L., Uzzell, V. J., Simoncelli, E. P., & Chichilnisky, E. J. (2005). Prediction
+and decoding of retinal ganglion cell responses with a probabilistic spiking model. Journal of
+Neuroscience , 25 (47), 11003-11013.
+
+Pillow, Jonathan W., Jonathon Shlens, Liam Paninski, Alexander Sher, Alan M. Litke, E. J.
+Chichilnisky, and Eero P. Simoncelli. "Spatio-temporal correlations and visual signalling in a
+complete neuronal population." Nature 454, no. 7207 (2008): 995-999.
+
+Gerwinn, Sebastian, Matthias Bethge, Jakob H. Macke, and Matthias Seeger. "Bayesian
+inference for spiking neuron models with a sparsity prior." In Advances in neural information
+processing systems , pp. 529-536. 2008.
+
+Zhao, M., & Iyengar, S. (2010). Nonconvergence in logistic and poisson models for neural
+spiking. Neural Computation, 22(5), 1231–1244.
+
+Macke, Jakob H., Lars Buesing, John P. Cunningham, M. Yu Byron, Krishna V. Shenoy, and
+Maneesh Sahani. "Empirical models of spiking in neural populations." In Advances in neural
+information processing systems , pp. 1350-1358. 2011.
+
+Park, I. M., & Pillow, J. W. (2011). Bayesian Spike Triggered Covariance Analysis. Advances in
+Neural Information Processing Systems (NeurIPS).
+
+Park, M., & Pillow, J. W. (2011). Receptive Field Inference with Localized Priors. PLoS
+Computational Biology, 7(10), e1002219+.
+
+Pillow, Jonathan W., and James Scott. "Fully Bayesian inference for neural models with
+negative-binomial spiking." In Advances in neural information processing systems , pp.
+1898-1906. 2012.
+
+Stevenson, I. H., London, B. M., Oby, E. R., Sachs, N. A., Reimer, J., Englitz, B., David, S. V.,
+Shamma, S. A., Blanche, T. J., Mizuseki, K., Zandvakili, A., Hatsopoulos, N. G., Miller, L. E., &
+Kording, K. P. (2012). Functional connectivity and tuning curves in populations of simultaneously
+recorded neurons. PLoS Computational Biology, 8(11), e1002775.
+
+Vidne, Michael, Yashar Ahmadian, Jonathon Shlens, Jonathan W. Pillow, Jayant Kulkarni, Alan
+M. Litke, E. J. Chichilnisky, Eero Simoncelli, and Liam Paninski. "Modeling the impact of common
[…]
