+++ tutorials/W1D5_DimensionalityReduction/further_reading.md
@@ -1,0 +1,34 @@
+# Dimensionality Reduction
+## Suggested further readings
+&nbsp;
+
+### Principal component analysis (PCA):
+
+Shlens (2014) A tutorial on principal component analysis. arXiv https://arxiv.org/abs/1404.1100
+
+### Other linear dimensionality reductions:
+
+Hyvarinen & Oja (2000) Independent component analysis: algorithms and applications. https://www.sciencedirect.com/science/article/abs/pii/S0893608000000265
+
+Gillis (2014) The why and how of nonnegative matrix factorization. https://arxiv.org/abs/1401.5226
+
+### Nonlinear dimensionality reduction:
+
+Wattenberg et al. (2016) How to Use t-SNE Effectively Distill. http://doi.org/10.23915/distill.00002
+
+Understanding UMAP. https://pair-code.github.io/understanding-umap/
+
+### Dimensionality reduction in systems neuroscience:
+
+Cunningham & Yu (2014) Dimensionality reduction for large-scale neural recordings. Nature Neuroscience https://www.nature.com/articles/nn.3776
+
+### Brain-computer interface work shown in Outro:
+
+Golub et al. (2016) Brain–computer interfaces for dissecting cognitive processes
+underlying sensorimotor control. Current Opinion in Neurobiology https://www.sciencedirect.com/science/article/abs/pii/S095943881500183X
+
+Sadtler et al. (2014) Neural constraints on learning. Nature https://www.nature.com/articles/nature13665
+
+Golub et al. (2018) Learning by neural reassociation. Nature Neuroscience https://www.nature.com/articles/s41593-018-0095-3
+
+&bnsp
