+++ tutorials/W1D1_ModelTypes/further_reading.md
@@ -1,0 +1,64 @@
+# Model Types
+## Suggested further readings 
+
+&nbsp;
+
+Bassett, D. S., Zurn, P., & Gold, J. I. (2018). On the nature and use of models in network neuroscience. Nature Reviews. Neuroscience. https://doi.org/10.1038/s41583-018-0038-8 
+
+Bennett, M. R., & Hacker, P. M. S. (2003). Philosophical Foundations of Neuroscience. Wiley-Blackwell. 
+Blohm, Gunnar, Konrad P. Kording, and Paul R. Schrater. 2020. “A How-to-Model Guide for Neuroscience.” eNeuro 7 (1). https://doi.org/10.1523/ENEURO.0352-19.2019. 
+
+Burgess, J. (1998). Occam’s razor and scientific method. In The Philosophy of Mathematics Today (pp. 195–214). Clarendon Press, Oxford. 
+
+Chandrasekhar, S. (2013). Truth and Beauty: Aesthetics and Motivations in Science. University of Chicago Press. 
+Chater, N., & Oaksford, M. (1999). Ten years of the rational analysis of cognition. Trends in Cognitive Sciences, 3(2), 57–65. 
+
+Churchland, P. S., & Sejnowski, T. J. (1990). Neural Representation and Neural Computation. 
+Philosophical Perspectives, Vol. 4, p. 343. https://doi.org/10.2307/2214198 
+
+Churchland, P. S., & Sejnowski, T. J. (1988). Perspectives on cognitive neuroscience. Science, 242(4879), 741–745. 
+Cichy, Radoslaw M., and Daniel Kaiser. 2019. “Deep Neural Networks as Scientific Models.” 
+Trends in Cognitive Sciences. https://doi.org/10.1016/j.tics.2019.01.009. 
+
+Dayan, P. (2005). Theoretical Neuroscience: Computational And Mathematical Modeling of Neural Systems. Mit Press. 
+Feldman, J. (2016). The simplicity principle in perception and cognition. Wiley Interdisciplinary Reviews. Cognitive Science, 7(5), 330–340. 
+
+Gillett, C. (2016). Reduction and Emergence in Science and Philosophy. Cambridge University Press. 
+Goldstein, R. E. (2018). Are theoretical results “Results”? eLife, Vol. 7. 
+https://doi.org/10.7554/elife.40018 
+
+Jonas, E., & Kording, K. P. (2017). Could a Neuroscientist Understand a Microprocessor? PLoS Computational Biology, 13(1), e1005268. 
+
+Josephson, J. R., & Josephson, S. G. (Eds.). (1994). Abductive inference. Cambridge: Cambridge University Press. 
+
+Kaplan, David Michael. 2011. “Explanation and Description in Computational Neuroscience.” 
+	Synthese. https://doi.org/10.1007/s11229-011-9970-0. 
+    
+Kording, Konrad, Gunnar Blohm, Paul Schrater, and Kendrick Kay. n.d. “Appreciating Diversity of Goals in Computational Neuroscience.” https://doi.org/10.31219/osf.io/3vy69. 
+
+Lee, Michael David, Amy H. Criss, Berna Devezer, Chris Donkin, Alexander Etz, Fabio P. Leite, Dora Matzke, et al. n.d. “Robust Modeling in Cognitive Science.” https://doi.org/10.31234/osf.io/dmfhk. 
+
+Lombrozo, T. (2012). Explanation and Abductive Inference. Oxford Handbooks Online. https://doi.org/10.1093/oxfordhb/9780199734689.013.0014 
+
+Marr, D., & Poggio, T. (1976). From Understanding Computation to Understanding Neural 
[…]
