+++ tutorials/W3D4_ReinforcementLearning/further_reading.md
@@ -1,0 +1,39 @@
+# Reinforcement Learning
+## Suggested further readings
+&nbsp;
+
+### Overview
+
+Sutton, R. S., & Barto, A. G. (2018). Reinforcement learning: An introduction. MIT press.
+
+### Links to neuroscience
+Schultz, W., Dayan, P., & Montague, P. R. (1997). A neural substrate of prediction and reward.
+Science, 275(5306), 1593-1599.
+
+Daw, N. D., Niv, Y., & Dayan, P. (2005). Uncertainty-based competition between prefrontal and
+dorsolateral striatal systems for behavioral control. Nature neuroscience, 8(12), 1704-1711.
+
+Dayan, P., & Niv, Y. (2008). Reinforcement learning: the good, the bad and the ugly. Current
+opinion in neurobiology, 18(2), 185-196.
+
+Wang, J. X., Kurth-Nelson, Z., Kumaran, D., Tirumala, D., Soyer, H., Leibo, J. Z., ... & Botvinick,
+M. (2018). Prefrontal cortex as a meta-reinforcement learning system. Nature neuroscience,
+21(6), 860-868.
+
+Mattar, M. G., & Daw, N. D. (2018). Prioritized memory access explains planning and
+hippocampal replay. Nature neuroscience, 21(11), 1609-1617.
+
+### State of the art
+Dabney, W., Kurth-Nelson, Z., Uchida, N., Starkweather, C. K., Hassabis, D., Munos, R., &
+Botvinick, M. (2020). A distributional code for value in dopamine-based reinforcement learning.
+Nature, 577(7792), 671-675.
+
+Mnih, V., Kavukcuoglu, K., Silver, D., Rusu, A. A., Veness, J., Bellemare, M. G., ... & Petersen,
+S. (2015). Human-level control through deep reinforcement learning. nature, 518(7540),
+529-533.
+
+Silver, D., Huang, A., Maddison, C. J., Guez, A., Sifre, L., Van Den Driessche, G., ... &
+Dieleman, S. (2016). Mastering the game of Go with deep neural networks and tree search.
+nature, 529(7587), 484-489.
+
+&nbsp;
