+++ tutorials/W2D3_RealNeurons/further_reading.md
@@ -1,0 +1,84 @@
+# Biological Neuron Models
+## Suggested further readings
+&nbsp;
+
+### Text book:
+
+Dayan P, Abbott LF. Theoretical neuroscience: computational and mathematical modeling of neural
+systems. Journal of Cognitive Neuroscience. 2003 Jan;15(1):154-5. Chapter 5
+
+Gerstner W, Kistler WM, Naud R, Paninski L. Neuronal dynamics: From single neurons to networks
+and models of cognition. Cambridge University Press; 2014 Jul 24. Chapter 1
+Research papers
+
+### Hodgkin-Huxley neuron model (how spike is generated):
+
+Hodgkin AL, Huxley AF. A quantitative description of membrane current and its application to
+conduction and excitation in nerve. The Journal of physiology. 1952 Aug 28;117(4):500.
+
+Hodgkin AL, Huxley AF, Katz B. Measurement of current-voltage relations in the membrane of the
+giant axon of Loligo. The Journal of physiology. 1952 Apr 28;116(4):424.
+
+### But what is a spike
+
+FitzHugh R. Mathematical models of threshold phenomena in the nerve membrane. The bulletin of
+mathematical biophysics. 1955 Dec 1;17(4):257-78.
+
+Izhikevich EM, FitzHugh R. Fitzhugh-nagumo model. Scholarpedia. 2006 Sep 23;1(9):1349.
+
+### Point neuron model
+
+Brunel N, Van Rossum MC. Lapicque’s 1907 paper: from frogs to integrate-and-fire. Biological
+cybernetics. 2007 Dec 1;97(5-6):337-9.
+
+Gerstner W, Naud R. How good are neuron models?. Science. 2009 Oct 16;326(5951):379-80.
+
+Jolivet R, Kobayashi R, Rauch A, Naud R, Shinomoto S, Gerstner W. A benchmark test for a
+quantitative assessment of simple neuron models. Journal of neuroscience methods. 2008 Apr
+30;169(2):417-24.
+
+Jolivet R, Lewis TJ, Gerstner W. Generalized integrate-and-fire models of neuronal activity
+approximate spike trains of a detailed model to a high degree of accuracy. Journal of
+neurophysiology. 2004 Aug;92(2):959-76.
+
+Brunel N, Hakim V, Richardson MJ. Single neuron dynamics and computation. Current opinion in
+neurobiology. 2014 Apr 1;25:149-55.
+
+### Extending simplified neuron models
+
+Poirazi P, Brannon T, Mel BW. Pyramidal neuron as two-layer neural network. Neuron. 2003 Mar
+27;37(6):989-99.
+
+Larkum ME, Nevian T, Sandler M, Polsky A, Schiller J. Synaptic integration in tuft dendrites of layer
+5 pyramidal neurons: a new unifying principle. Science. 2009 Aug 7;325(5941):756-60.
+
+### Synapse models
+
+Kuhn A, Aertsen A, Rotter S. Neuronal integration of synaptic input in the fluctuation-driven regime.
+Journal of Neuroscience. 2004 Mar 10;24(10):2345-56.
+
+### Short-term dynamics of synapses
+
+Stevens CF, Wang Y. Facilitation and depression at single central synapses. Neuron. 1995 Apr
+1;14(4):795-802.
+
+Markram H, Tsodyks M. Redistribution of synaptic efficacy between neocortical pyramidal neurons.
+Nature. 1996 Aug;382(6594):807-10.
+
+Markram H, Wang Y, Tsodyks M. Differential signaling via the same axon of neocortical pyramidal
[…]

[…]
