### 2021-06-22 fa52be9
Rename W1D4 to Generalized Linear Models
