+++ tutorials/README.md
@@ -2,7 +2,7 @@
 
 <!-- DO NOT EDIT THIS FILE. IT IS AUTO-GENERATED BY A FRIENDLY ROBOT -->
 
-[W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---machine-learning) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---real-neurons) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
+[W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---generalized-linear-models) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---real-neurons) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
 
 *Warning:* The 'render with NBViewer' buttons may show outdated content.
 
@@ -107,7 +107,7 @@ Slides: [Intro](https://mfr.ca-1.osf.io/render?url=https://osf.io/sqcvz/?direct%
 
 
 
-## W1D4 - Machine Learning
+## W1D4 - Generalized Linear Models
 
 [YouTube Playlist](https://www.youtube.com/playlist?list=PLkBQOLLbi18NQvqJOvyUxeaOwuEICnWB6)
 
@@ -115,8 +115,8 @@ Slides: [Intro](https://mfr.ca-1.osf.io/render?url=https://osf.io/qxfz9/?direct%
 
 |   | Run | View |
 | - | --- | ---- |
-| Tutorial 1 | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/tutorials/W1D4_MachineLearning/student/W1D4_Tutorial1.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/tutorials/W1D4_MachineLearning/student/W1D4_Tutorial1.ipynb?flush_cache=true) |
-| Tutorial 2 | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/tutorials/W1D4_MachineLearning/student/W1D4_Tutorial2.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/tutorials/W1D4_MachineLearning/student/W1D4_Tutorial2.ipynb?flush_cache=true) |
[…]
