+++ tutorials/materials.yml
@@ -85,7 +85,7 @@
 - day: W1D4
   category: Machine Learning
   intro: https://www.youtube.com/watch?v=m1w7oywzwpA
-  name: Machine Learning
+  name: Generalized Linear Models
   outro: https://www.youtube.com/watch?v=NXVG9ORBYXQ
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18NQvqJOvyUxeaOwuEICnWB6
   qa:
@@ -165,7 +165,7 @@
 - day: W2D3
   category: Dynamical Systems
   intro: https://www.youtube.com/watch?v=MAOOPv3whZ0
-  name: Real Neurons
+  name: Biological Neuron Models
   outro: https://www.youtube.com/watch?v=pMxb9vA-_4U
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18MCEdPJQ7gdnqP-Z0Tkmcjy
   qa:
