### 2021-06-23 16e23d6
Add instructions file to run notebooks on Kaggle
