+++ organization/Tutorial_kaggle.md
@@ -21,5 +21,5 @@
 
 ![Step 6.1](../tutorials/static/kaggle_step6_1.png)
 
-Once your phone number has been verified, on the left sidebar under **Settings** panel, select **Accelerator** to **GPU** in the drop down menu.
+Once your phone number has been verified, on the left sidebar under **Settings** panel, select **Accelerator** to **GPU** from **None** using the drop down menu.
 ![Step 6.2](../tutorials/static/kaggle_step6_2.png)
