### 2021-06-23 b2f1ac5
Add kaggle instructions (#594)
* Add instructions file to run notebooks on Kaggle

* Add images and include additional steps

* Change font

* Clarify step 6
