+++ organization/Tutorial_kaggle.md
@@ -1,0 +1,25 @@
+# Steps to run notebooks on Kaggle with GPU support
+
+## STEPS:
+1. First, [sign in](https://www.kaggle.com/account/login) to your Kaggle account. If you don't have a Kaggle account, create one by registering for it by clicking [here](https://www.kaggle.com/account/login?phase=startRegisterTab&returnUrl=%2F). 
+![Step 1](../tutorials/static/kaggle_step1.png)
+
+2. Once you're logged into your account, head over to the left sidebar menu and click on the **Code** section. 
+![Step 2](../tutorials/static/kaggle_step2.png)
+
+3. Then, click on the **New Notebook** button on the page.
+![Step 3](../tutorials/static/kaggle_step3.png)
+
+4. In the **File** menu bar, select **Upload Notebook** option.
+![Step 4](../tutorials/static/kaggle_step4.png)
+
+5. Next, select **Upload via external URL** option, and input the url of the notebook. For instance, if you wish you run [W2D1_Tutorial.ipynb](https://github.com/NeuromatchAcademy/course-content/blob/master/tutorials/W2D1_DeepLearning/W2D1_Tutorial1.ipynb) get its corresponding link (https://github.com/NeuromatchAcademy/course-content/blob/master/tutorials/W2D1_DeepLearning/W2D1_Tutorial1.ipynb), and then select **Upload** button to import the notebook on Kaggle.
+![Step 5](../tutorials/static/kaggle_step5.png)
+
+6. The last and final step but important **step** about how to leverage **GPU** accelerator in the kaggle kernel. To get access to the GPU, first time users have get their phone numbers verified. In order to have GPU access, on the left sidebar of the notebook under the settings panel, go to **Get phone verified** link and enter your credentials.
+   
+
+![Step 6.1](../tutorials/static/kaggle_step6_1.png)
+
+Once your phone number has been verified, on the left sidebar under **Settings** panel, select **Accelerator** to **GPU** from **None** using the drop down menu.
+![Step 6.2](../tutorials/static/kaggle_step6_2.png)
