+++ tutorials/materials.yml
@@ -21,8 +21,10 @@
 - day: W1D1
   category: Intro to Modeling
   intro: https://www.youtube.com/watch?v=KxldhMR5PxA
+  intro_bilibili: https://www.bilibili.com/video/BV1HT4y1E7U4/
   name: Model Types
   outro: https://www.youtube.com/watch?v=KZQXfQL1SH4
+  outro_bilibili: https://www.bilibili.com/video/BV1vv411i7SG/
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18ObAiSOZ42YBwOQIKNvspeI
   qa:
   - https://www.youtube.com/watch?v=FwsKmunwzUA
@@ -43,8 +45,10 @@
 - day: W1D2
   category: Intro to Modeling
   intro: https://www.youtube.com/watch?v=65HhPd0kG2k
+  intro_bilibili: https://www.bilibili.com/video/BV1Jk4y1B7cz/
   name: Modeling Practice
   outro: https://www.youtube.com/watch?v=JSTjK6PrOgY
+  outro_bilibili: https://www.bilibili.com/video/BV1mT4y177qN
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PxZse6mPHnxviOnp5bMs-g
   qa:
   - https://www.youtube.com/watch?v=j8xsEOg8F_Q
@@ -61,12 +65,14 @@
     title: Reading
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/v5emn/?direct%26mode=render%26action=download%26mode=render
     title: TA
-  tutorials: 2
+  tutorials: 1
 - day: W1D3
   category: Machine Learning
   intro: https://www.youtube.com/watch?v=yTQNqYCvoZ0
+  intro_bilibili: https://www.bilibili.com/video/BV1Wf4y1R75n
   name: Model Fitting
   outro: https://www.youtube.com/watch?v=cUAUL3efqpY
+  outro_bilibili: https://www.bilibili.com/video/BV1vK4y1s7CR
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18MqTMnc_mMiuawnczEa8xm9
   qa:
   - https://www.youtube.com/watch?v=lYLa6zgqByw
@@ -85,8 +91,10 @@
 - day: W1D4
   category: Machine Learning
   intro: https://www.youtube.com/watch?v=m1w7oywzwpA
+  intro_bilibili: https://www.bilibili.com/video/BV1BK411H7ie/
   name: Generalized Linear Models
   outro: https://www.youtube.com/watch?v=NXVG9ORBYXQ
+  outro_bilibili: https://www.bilibili.com/video/BV1d5411Y7mN
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18NQvqJOvyUxeaOwuEICnWB6
   qa:
   - https://www.youtube.com/watch?v=98FRcmPQduQ
@@ -105,8 +113,10 @@
 - day: W1D5
   category: Machine Learning
   intro: https://www.youtube.com/watch?v=zeBFyRaoVnQ
+  intro_bilibili: https://www.bilibili.com/video/BV1Cf4y117b6/
   name: Dimensionality Reduction
   outro: https://www.youtube.com/watch?v=Vaut2BJlv10
+  outro_bilibili: https://www.bilibili.com/video/BV1pi4y1V7tK/
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18N0E0Bx-ZueacF2oRr5eEpU
   qa:
   - https://www.youtube.com/watch?v=WsKeH6UmZLg
@@ -125,8 +135,10 @@
[…]
