### 2021-06-23 1d4c304
Add new materials (#593)
* Adding code for open in colab in new tab (#591)

* Add day overviews

* Code to hide cells in jupyterbook (#592)

* Adding code for open in colab in new tab

* Adding code to hide colab hidden cells in jupyterbook

* Add further readings

* Add further readings

* Rename W1D4 to Generalized Linear Models

* Process tutorial notebooks

* Rename W2D3 to Biological Neuron Models

* Process tutorial notebooks

* Rename W2D3 to Biological Neuron Models

* Process tutorial notebooks

* Change further reading headers

* Change further reading headers

* Process tutorial not
[…]
