+++ tutorials/W2D2_LinearSystems/further_reading.md
@@ -1,0 +1,134 @@
+# Suggested further readings
+
+### Further reading on linear dynamical systems models in neuroscience:
+
+Billeh, Y. N., Cai, B., Gratiy, S. L., Dai, K., Iyer, R., Gouwens, N. W., Abbasi-Asl, R., Jia, X., Siegle, J. H.,
+Olsen, S. R., Koch, C., Mihalas, S., & Arkhipov, A. (2020). Systematic Integration of Structural and
+Functional Data into Multi-scale Models of Mouse Primary Visual Cortex. Neuron, 106(3), 388-403.e18.
+https://doi.org/10.1016/j.neuron.2020.01.040
+
+Brunton, B. W., Botvinick, M. M., & Brody, C. D. (2013). Rats and Humans Can Optimally Accumulate
+Evidence for Decision-Making. Science, 340(6128), 95–98. https://doi.org/10.1126/science.1233912
+
+Brunton, B. W., Johnson, L. A., Ojemann, J. G., & Kutz, J. N. (2016). Extracting spatial-temporal coherent
+patterns in large-scale neural recordings using dynamic mode decomposition. Journal of neuroscience
+methods, 258, 1–15. https://doi.org/10.1016/j.jneumeth.2015.10.010
+
+Costa, A. C., Ahamed, T., & Stephens, G. J. (2019). Adaptive, locally linear models of complex dynamics.
+Proceedings of the National Academy of Sciences, 116(5), 1501–1510. https://doi.org/10.1073/pnas.1813476116
+
+Gilson, M., Burkitt, A. N., Grayden, D. B., Thomas, D. A., & van Hemmen, J. L. (2009). Emergence of
+network structure due to spike-timing-dependent plasticity in recurrent neuronal networks. I. Input
+selectivity--strengthening correlated input pathways. Biological cybernetics, 101(2), 81–102. https://doi.org/10.1007/s00422-009-0319-4
+
+Harris, K. D., Aravkin, A., Rao, R., & Brunton, B. W. (2020). Time-varying autoregression with low rank
+tensors. ArXiv:1905.08389 Cs, Stat. http://arxiv.org/abs/1905.08389
+
+Hodgkin, A. L., & Huxley, A. F. (1952). A quantitative description of membrane current and its application
+to conduction and excitation in nerve. The Journal of physiology, 117(4), 500–544. https://doi.org/10.1113/jphysiol.1952.sp004764
+
+Hu, Y., Brunton, S. L., Cain, N., Mihalas, S., Kutz, J. N., & Shea-Brown, E. (2018). Feedback through
+graph motifs relates structure and function in complex networks. Physical Review E, 98(6), 062312.
+https://doi.org/10.1103/physreve.98.062312
+
+Izhikevich, E.M. (2007). Dynamical Systems in Neuroscience: The Geometry of Excitability and Bursting.
+MIT Press.
+
+Linderman, S. W., Miller, A. C., Adams, R. P., Blei, D. M., Paninski, L., & Johnson, M. J. (2016).
+Recurrent switching linear dynamical systems. ArXiv:1610.08466 Stat. http://arxiv.org/abs/1610.08466
+
+Mante, V., Sussillo, D., Shenoy, K. V., & Newsome, W. T. (2013). Context-dependent computation by
+recurrent dynamics in prefrontal cortex. Nature, 503(7474), 78–84. https://doi.org/10.1038/nature12742
+
+Morrison, K., & Curto, C. (2018). Predicting neural network dynamics via graphical analysis.
+ArXiv:1804.01487 Math, q-Bio. http://arxiv.org/abs/1804.01487
+
[…]
