### 2021-06-23 bea5e7d
Fixed jupyterbook layout issues
