### 2021-06-23 333884b
Fixed jupyterbook layout issues with solution figures
Only edited the extract_solutions function in ci/process_notebooks.py by adding a
(newline) after the html solution image. Comment code below this change provides an alternatively solution using the "fixed-width" metadata tag for the cell; you also have to uncomment the code specifying the width/height args passed into the html  tag. The solution images this method generates are larger than ideal, but solves the layout issue too.
