### 2021-06-23 a902ff8
Remove intro text files
