+++ tutorials/W1D5_DimensionalityReduction/intro_text.md
@@ -1,3 +1,0 @@
-# Dimensionality Reduction
-
-Today, we'll learn all about dimensionality reduction.
