+++ tutorials/materials.yml
@@ -223,8 +223,10 @@
 - day: W3D1
   category: Stochastic Processes
   intro: https://www.youtube.com/watch?v=w2GH1-Iqods
+  intro_bilibili: ""
   name: Bayesian Decisions
   outro: https://www.youtube.com/watch?v=eOEyDjhlHwA
+  outro_bilibili: ""
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PTkIw6eezzVM0UmrBzwMok
   qa:
   - https://www.youtube.com/watch?v=CVXe6fJ9xdI
