### 2021-06-23 3bf4657
Initial auto generation of intro/outro files
