+++ tutorials/materials.yml
@@ -138,7 +138,9 @@
   intro_bilibili: https://www.bilibili.com/video/BV1sk4y1B7Ej
   name: Deep Learning
   outro: https://www.youtube.com/watch?v=pzA1GpxodnM
+  outro_2: https://www.youtube.com/watch?v=nWlgIclpyt4
   outro_bilibili: https://www.bilibili.com/video/BV1M54y1B7hs
+  outro_bilibili_2: https://www.bilibili.com/video/BV1GT4y1j7aQ
   playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18PZ2uw0p7G4EkzjzqP8l0Eg
   qa:
   - https://www.youtube.com/watch?v=c7DftQoCcf8
