### 2021-06-23 b883e6f
Delete rest of intro text files
