+++ tutorials/W2D4_DynamicNetworks/intro_text.md
@@ -1,7 +1,0 @@
-# Dynamic Networks
-
-Today you will learn about dynamical properties of biological neuronal networks. At the start, Nicolas Brunel will introduce you to the building blocks of biological neuronal networks and various types of activity that may arise in such networks. He will start with a network of spiking neurons and zoom out to firing rate-based models. In the tutorials we will focus on firing rate-based network models only. In the first tutorial Juliana Georgieva will develop the notion of firing rate-based networks and introduce the idea of fixed-points and stability of a population of neurons. In the second tutorial she will extend this idea to a two-population network model (the Wilson-Cowan model). In this tutorial, you will learn how you can use phase plane analysis to comment about the dynamics that arises because of excitatory and inhibitory interactions e.g. oscillations and attractors etc. Finally, in the outro lecture Ken Miller will use the firing rate-based models to expose three very interesting modes of biological neuronal networks i.e. inhibition stabilized network, balanced amplification and stabilized supra linear networks. He will show how these modes can provide mechanisms underlying several biological phenomena such as surround suppression in the visual cortex.
- 
-The networks we will use today are built on what you have learned about Linear Systems (W2D2) and Biological Neurons (W2D3). An understanding of the stability and dynamical states of the neuronal network will also help you appreciate the problem of causality (W3D5) better. Moreover, a theoretical understanding of the network activity dynamics will help you develop better statistical controls for data analysis and model fitting – something you will encounter throughout the course and in your projects.
- 
-Information processing the brain is manifested in terms of changes in network activity. Large scale changes in the network activity e.g. transitions from non-oscillatory state to oscillatory state, do happen in the brain and have huge impact on computations. What are the mechanisms that lead to change in the dynamical state of network activity? What determines the stability of network activity state? How is a balanced neuronal network able to amplify small changes while rejecting average value? How multiple inputs may interact to shape the response? Such questions routinely arise in neuroscience. The contents of the tutorial today should give you a good starting point to start addressing such questions and identify mechanisms and causes of network activity changes.
