### 2021-06-24 75a9273
Delete old ways of storing file info
