+++ tutorials/README.md
@@ -2,13 +2,14 @@
 
 <!-- DO NOT EDIT THIS FILE. IT IS AUTO-GENERATED BY A FRIENDLY ROBOT -->
 
-[W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---generalized-linear-models) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---biological-neuron-models) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
+[W0D1](#w0d1---python-workshop1) | [W0D2](#w0d2---python-workshop2) | [W0D3](#w0d3---linear-algebra) | [W0D4](#w0d4---calculus) | [W0D5](#w0d5---statistics) | [W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---generalized-linear-models) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---biological-neuron-models) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
 
 *Warning:* The 'render with NBViewer' buttons may show outdated content.
 
 ## W0D1 - Python Workshop1
 
 [YouTube Playlist](https://www.youtube.com/playlist?list=PLkBQOLLbi18M8aIPAL3n2q7D_AbbuPPtM)
+
 |   | Run | View |
 | - | --- | ---- |
 | Tutorial 1 | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/NeuromatchAcademy/course-content/blob/master/tutorials/W0D1_PythonWorkshop1/student/W0D1_Tutorial1.ipynb) | [![View the notebook](https://img.shields.io/badge/render-nbviewer-orange.svg)](https://nbviewer.jupyter.org/github/NeuromatchAcademy/course-content/blob/master/tutorials/W0D1_PythonWorkshop1/student/W0D1_Tutorial1.ipynb?flush_cache=true) |
@@ -19,6 +20,7 @@
 ## W0D2 - Python Workshop2
 
 [YouTube Playlist](https://www.youtube.com/playlist?list=PLkBQOLLbi18Oi40BY9HyzVlJFHxhph6Gy)
+
 |   | Run | View |
 | - | --- | ---- |
[…]
