### 2021-06-24 48e9826
Add playlists to pre-reqs materials
