+++ tutorials/materials.yml
@@ -1,22 +1,27 @@
 - day: W0D1
   category: Pre-reqs Refresher
   name: Python Workshop 1
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18M8aIPAL3n2q7D_AbbuPPtM
   tutorials: 1
 - day: W0D2
   category: Pre-reqs Refresher
   name: Python Workshop 2
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18Oi40BY9HyzVlJFHxhph6Gy
   tutorials: 1
 - day: W0D3
   category: Pre-reqs Refresher
   name: Linear Algebra
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18Oz3lb0fo06k0cAWk54rsoE
   tutorials: 3
 - day: W0D4
   category: Pre-reqs Refresher
   name: Calculus
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18NTiaHYRylxN0gSm1PlQ15N
   tutorials: 3
 - day: W0D5
   category: Pre-reqs Refresher
   name: Statistics
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18MDDH09-kvRoU8XANuCJvGg
   tutorials: 2
 - day: W1D1
   category: Intro to Modeling
