### 2021-06-24 c85cf16
Change workflow to add yaml install
