### 2021-06-24 430a0a8
Change workflow to add yaml install
