+++ .github/workflows/tutorial-readme-pr.yaml
@@ -24,7 +24,7 @@ jobs:
 
       - name: Install dependencies
         run: |
-          pip install -r requirements.txt
+          pip install pyyaml
 
       - name: Update READMEs
         run: python ci/generate_tutorial_readmes.py
