+++ .github/workflows/tutorial-readme-pr.yaml
@@ -2,6 +2,8 @@ name: tutorial-readme-pr
 on: 
   pull_request:
     branches: master
+  paths:
+    - tutorials/materials.yaml
 
 jobs:
 
