+++ .github/workflows/notebook-pr.yaml
@@ -2,7 +2,9 @@ name: notebook-pr
 on: 
   pull_request:
     branches: master
-    paths: '**/*.ipynb'
+    paths: 
+      - '**/*.ipynb'
+      - ci/process_notebooks.py
 
 env:
   NB_KERNEL: python
