### 2021-06-24 71b88c4
Update config to execute notebooks, have favicon, and link to git
