### 2021-06-24 56e0d76
Update hiding of cells in jupyterbook
