### 2021-06-24 b9f6f17
Update hiding cells in jupyterbook
