### 2021-06-24 7480233
Update hiding cells in jupyterbook
