### 2021-06-24 523732c
force checkout to discard temp changes made to notebooks for book generation
