+++ .github/workflows/publish-book.yml
@@ -37,7 +37,7 @@ jobs:
         run: |
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
-          ghp-import -n -m "Update course book" book/_build/html
+          ghp-import -n -c "compneuro.neuromatch.io" -m "Update course book" book/_build/html
           git checkout -f gh-pages
 
       - name: Publish to gh-pages
