### 2021-06-24 83ff902
Book/force checkout (#600)
* force checkout to discard temp changes made to notebooks for book generation

* need to add cname or it gets removed on deploy
