### 2021-06-24 64f1c24
Update course book
