+++ _sources/tutorials/W2D1_DeepLearning/outro_vid.md
@@ -1,7 +1,7 @@
 # Outro Video
 ## Video
 
-<iframe width="420" height="315" src="https://www.youtube.com/embed/nWlgIclpyt4"></iframe>
+<iframe width="420" height="315" src="https://www.youtube.com/embed/pzA1GpxodnM"></iframe>
 
 ## Slides
 <iframe src="https://mfr.ca-1.osf.io/render?url=https://osf.io/z5g93/?direct%26mode=render%26action=download%26mode=render", frameborder="0" width="960" height="569" allowfullscreen="true" mozallowfullscreen="true" webkitallowfullscreen="true"></iframe>
