+++ tutorials/W0D3_LinearAlgebra/intro_text.md
@@ -1,8 +1,0 @@
-# W0D3 - Linear Algebra
-
-Paragraph 1. A very brief paragraph overviewing the day
-
-Paragraph 2. High level overview of how this content fits in the field of comp neuro/course, how it builds on previous days, how it will connect to future days. This can be pretty broad and brief. 
-
-Paragraph 3. A couple examples of what sorts of research questions the content is useful for. Again, this is very basic and introductory so no details necessary. 
-
