+++ tutorials/W3D1_BayesianDecisions/intro_text.md
@@ -1,3 +1,0 @@
-# Bayesian Decisions
-
-Today, we'll be learning about Bayesian Decisions.

[…]
