### 2021-06-24 7c342d7
Update course book
