+++ _sources/tutorials/W2D2_LinearSystems/intro_text.md
@@ -1,3 +1,0 @@
-# Linear Systems
-
-Today, we'll be discussing linear systems.
