+++ _sources/tutorials/W3D4_ReinforcementLearning/intro_text.md
@@ -1,3 +1,0 @@
-# Reinforcement Learning
-
-Today, we'll learn all about reinforcement learning.
