### 2021-06-24 NeuromatchAcademy/course-content#601 @PaulScotti
JOSE: Versioning
The GitHub page indicates that this repository is currently under construction for the 2021 course (with a link to the 2020 archival materials in the README), but given that this paper is describing materials from the 2020 course then perhaps the associated release for this work should be the NMA2020 release (https://github.com/NeuromatchAcademy/course-content/releases/tag/NMA2020)? Currently the JOSE system appears to be interpreting the release version as “1” which I think this incorrect.

This is reviewer feedback associated with submission to JOSE: https://github.com/openjournals/jose-reviews/issues/118
