### 2021-06-24 NeuromatchAcademy/course-content#602 @PaulScotti
JOSE: References
Each dataset (Table 2) should be cited in the References. The current reference (van Viegen et al) should be changed to the now-published TICS paper (it is currently citing the arXiv preprint).

This is reviewer feedback associated with submission to JOSE: openjournals/jose-reviews#118
