### 2021-06-24 NeuromatchAcademy/course-content#603 @PaulScotti
JOSE: Formatting
I think there is a minor error in the pdf: “All lectures and tutorial are available online [permalink]” I do not see a link here  in the pdf.

This is reviewer feedback associated with submission to JOSE: openjournals/jose-reviews#118
