### 2021-06-24 #118 @PaulScotti
Overall I am a fan of Neuromatch Academy and impressed by the success of the online summer school to foster such accessible, extensive, and collaborative teaching in computational neuroscience supported by hundreds of volunteers. Thank you to everyone involved in making NMA such a success.

The GitHub repository generally looks to be very good, with clearly defined instructions and an organized structure. My below suggestions pertain to the JOSE paper itself and how it is linked to the GitHub repo. It explains why I left "Version" and "References" unchecked in my reviewer checklist.

1. The GitHub page indicates that this repository is currently under construction for the 2021 course (with a link to the 2020 archival materials in the README), but given that this paper is describing materials from the 2020 course then perhaps the associated release for this work should be the NMA2020 release (https://github.com/NeuromatchAcademy/course-content/releases/tag/NMA2020)? Currently the JOSE system appears to be interpreting the release version as “1” which I think this incorrect.  https://github.com/NeuromatchAcademy/course-content/issues/601

2. Each dataset (Table 2) should be cited in the References. The current reference (van Viegen et al) should be changed to the now-published TICS paper (it is currently citing the arXiv preprint). https://github.com/NeuromatchAcademy/course-content/issues/602

3. I think there is a minor error in the pdf: “All lectures and tutorial are available online [permalink]” I do not see a link here  in the pdf. https://github.com/NeuromatchAcademy/course-content/issues/603
