### 2021-06-24 188a1d2
Fix hiding cells
