### 2021-06-24 b7c9235
extended & updated bibliography
