+++ bibliography.bib
@@ -1,15 +1,122 @@
 
-@article{van_viegen_neuromatch_2020,
-	title = {Neuromatch {Academy}: {Teaching} {Computational} {Neuroscience} with global accessibility},
+@article{viegen2021,
+	title = {Neuromatch {Academy}: {Teaching} {Computational} {Neuroscience} with {Global} {Accessibility}},
+	volume = {25},
+	issn = {1364-6613, 1879-307X},
 	shorttitle = {Neuromatch {Academy}},
-	url = {http://arxiv.org/abs/2012.08973},
-	abstract = {Neuromatch Academy designed and ran a fully online 3-week Computational Neuroscience summer school for 1757 students with 191 teaching assistants working in virtual inverted (or flipped) classrooms and on small group projects. Fourteen languages, active community management, and low cost allowed for an unprecedented level of inclusivity and universal accessibility.},
-	urldate = {2021-02-22},
-	journal = {arXiv:2012.08973 [q-bio]},
-	author = "{van Viegen}, Tara and Akrami, Athena and Bonnen, Kate and DeWitt, Eric and Hyafil, Alexandre and Ledmyr, Helena and Lindsay, Grace W. and Mineault, Patrick and Murray, John D. and Pitkow, Xaq and Puce, Aina and Sedigh-Sarvestani, Madineh and Stringer, Carsen and Achakulvisut, Titipat and Alikarami, Elnaz and Atay, Melvin Selim and Batty, Eleanor and Erlich, Jeffrey C. and Galbraith, Byron V. and Guo, Yueqi and Juavinett, Ashley L. and Krause, Matthew R. and Li, Songting and Pachitariu, Marius and Straley, Elizabeth and Valeriani, Davide and Vaughan, Emma and Vaziri-Pashkam, Maryam and Waskom, Michael L. and Blohm, Gunnar and Kording, Konrad and Schrater, Paul and Wyble, Brad and Escola, Sean and Peters, Megan A. K.",
+	url = {https://www.cell.com/trends/cognitive-sciences/abstract/S1364-6613(21)00095-4},
+	doi = {10.1016/j.tics.2021.03.018},
+	language = {English},
+	number = {7},
+	urldate = {2021-06-24},
+	journal = {Trends in Cognitive Sciences},
+	author = "{van Viegen}, Tara and Akrami, Athena and Bonnen, Kathryn and DeWitt, Eric and Hyafil, Alexandre and Ledmyr, Helena and Lindsay, Grace W. and Mineault, Patrick and Murray, John D. and Pitkow, Xaq and Puce, Aina and Sedigh-Sarvestani, Madineh and Stringer, Carsen and Achakulvisut, Titipat and Alikarami, Elnaz and Atay, Melvin Selim and Batty, Eleanor and Erlich, Jeffrey C. and Galbraith, Byron V. and Guo, Yueqi and Juavinett, Ashley L. and Krause, Matthew R. and Li, Songting and Pachitariu, Marius and Straley, Elizabeth and Valeriani, Davide and Vaughan, Emma and Vaziri-Pashkam, Maryam and Waskom, Michael L. and Blohm, Gunnar and Kording, Konrad and Schrater, Paul and Wyble, Brad and Escola, Sean and Peters, Megan A. K.",
+	month = jul,
+	year = {2021},
+	pmid = {33994097},
+	keywords = {accessibility, computational neuroscience, data science, education, inclusivity, Python coding},
+	pages = {535--538}
+}
+
+@article{stringer2021,
+	title = {High-precision coding in visual cortex},
+	volume = {184},
+	issn = {0092-8674, 1097-4172},
[…]
