### 2021-06-24 f441472
added citations to updated/extended bibliography
also fixed a typo in one of the contributors initials
