+++ .github/workflows/publish-book.yml
@@ -25,6 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           pip install -r requirements.txt
+          pip install nbconvert
           pip install jupyter-book==0.10.2 ghp-import
 
       - name: Build the book
