### 2021-06-24 6c5198a
Roll back error hiding code
