### 2021-06-24 8abf321
Roll back error hiding
