### 2021-06-24 64afeca
Parse html for error outputs
