+++ ci/parse_html_for_errors.py
@@ -1,0 +1,31 @@
+import yaml
+from bs4 import BeautifulSoup
+
+def main():
+    with open('tutorials/materials.yml') as fh:
+        materials = yaml.load(fh, Loader=yaml.FullLoader)
+
+    html_directory = 'book/_build/html/'
+    for m in materials:
+        name = f"{m['day']}_{''.join(m['name'].split())}"
+
+        for i in range(m['tutorials']):
+            notebook_file_path = f"{html_directory}/tutorials/{name}/student/{m['day']}_Tutorial{i + 1}.html"
+
+            with open(notebook_file_path, 'r') as f:
+                contents = f.read()
+
+            parsed_html = BeautifulSoup(contents, features="html.parser")
+
+            mydivs = parsed_html.find_all("div", {"class": "cell_output docutils container"})
+
+            for div in mydivs:
+                if 'NotImplementedError' in str(div):
+                    div.decompose()
+
+            with open(notebook_file_path, 'w') as f:
+                f.write(str(parsed_html))
+
+
+if __name__ == '__main__':
+    main()
