+++ tutorials/W3D1_BayesianDecisions/chapter_title.md
@@ -1,0 +1,1 @@
+# Bayesian Decisions
