+++ .github/workflows/publish-book.yml
@@ -32,6 +32,7 @@ jobs:
           python ci/generate_book.py
           ln -s ../tutorials book/tutorials
           jupyter-book build book
+          python ci/parse_html_for_errors.py
 
       - name: Commit book
         run: |
