### 2021-06-24 258c27e
Add smart hiding of code and narrower videos
