+++ ci/generate_book.py
@@ -65,35 +65,12 @@ def main():
         yaml.dump(toc_list, fh)
 
 
-def convert_youtube_url_to_embed_url(url):
-    # TODO: Support Chinese equivalent URLs
-    url_key = url.split("=")[1]
-    return "https://www.youtube.com/embed/" + url_key
-
-
-def generate_page(info, directory, chapter, file_type):
-    if file_type.lower() in info:
-        chapter['sections'].append({'file': f'tutorials/{directory}/{file_type.lower()}_vid.md'})
-        with open(os.path.join("ci", "resources", "intro_outro_template.txt"), encoding="utf-8") as f:
-            template_string = f.read()
-            template = Template(template_string)
-            for slides in info['slides']:
-                if slides['title'] == file_type:
-                    prepared_template_string = template.render(type=file_type,
-                                                               video_source_url=convert_youtube_url_to_embed_url(
-                                                                   info[file_type.lower()]),
-                                                               slide_source_url=slides['link'])
-                    with open(os.path.join("tutorials", directory, file_type.lower() + "_vid.md"),
-                              "w+") as intro_vid_file:
-                        intro_vid_file.write(prepared_template_string)
-    return chapter
-
-
 def pre_process_notebook(file_path):
     try:
         with open(file_path, encoding="utf-8") as read_notebook:
             content = json.load(read_notebook)
         pre_processed_content = open_in_colab_new_tab(content)
+        pre_processed_content = change_video_widths(pre_processed_content)
         pre_processed_content = link_hidden_cells(pre_processed_content)
         with open(file_path, "w", encoding="utf-8") as write_notebook:
             json.dump(pre_processed_content, write_notebook, indent=1, ensure_ascii=False)
@@ -111,22 +88,68 @@ def open_in_colab_new_tab(content):
     cells[0]['source'][0] = str(parsed_html)
     return content
 
-
 def link_hidden_cells(content):
     cells = content['cells']
-    for cell in cells:
+    updated_cells = cells.copy()
+
+    i_updated_cell = 0
+    for i_cell, cell in enumerate(cells):
+        updated_cell = updated_cells[i_updated_cell]
         if "source" not in cell:
             continue
         source = cell['source'][0]
-        if source.startswith("#@title") or source.startswith("# @title"):
+
+        if source.startswith("#") and '@title' not in source:
+            header_level = source.count('#')
+
+        if '@title' in source or '@markdown' in source:
             if 'metadata' not in cell:
-                cell['metadata'] = {}
+                updated_cell['metadata'] = {}
             if 'tags' not in cell['metadata']:
-                cell['metadata']['tags'] = []
-            if "hide-input" not in cell['metadata']['tags']:
-                cell['metadata']['tags'].append("hide-input")
+                updated_cell['metadata']['tags'] = []
+
+            # Check if cell is video one
+            if 'YouTubeVideo' in ''.join(cell['source']) or 'IFrame' in ''.join(cell['source']):
+                if "remove-input" not in cell['metadata']['tags']:
+                    updated_cell['metadata']['tags'].append("remove-input")
+            else:
+                if "hide-input" not in cell['metadata']['tags']:
+                    updated_cell['metadata']['tags'].append("hide-input")
+
+            # If header is lost, create one in markdown
+            if '@title' in source:
+                if source.split('@title')[1] != '':
+                    header_cell = {
+                        'cell_type': 'markdown',
+                        'metadata': {},
+                        'source': ['#'*(header_level + 1) + ' ' + source.split('@title')[1]]}
+                    updated_cells.insert(i_updated_cell, header_cell)
+                    i_updated_cell += 1
+        i_updated_cell += 1
+
[…]
