### 2021-06-24 bedbcb7
Add parsing of html to hide code errors
