### 2021-06-24 1f6f73a
Rename Bonus Deep Learning 2 to autoencoders
