+++ tutorials/materials.yml
@@ -161,6 +161,24 @@
   - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/w6dfq/?direct%26mode=render%26action=download%26mode=render
     title: Reading
   tutorials: 3
+- day: Bonus
+  category: Machine Learning
+  intro: https://www.youtube.com/watch?v=VwSnDJZekQ4
+  intro_bilibili: https://www.bilibili.com/video/BV1LC4y187Pv/
+  name: Autoencoders
+  outro: https://www.youtube.com/watch?v=ksgsCy9EaR4
+  outro_bilibili: https://www.bilibili.com/video/BV1Df4y1X7jh/
+  playlist: https://www.youtube.com/playlist?list=PLkBQOLLbi18Ojl1CV8W00JZ0C0hivBxw1
+  slides:
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/pdnsg/?direct%26mode=render%26action=download%26mode=render
+    title: Intro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/4quda/?direct%26mode=render%26action=download%26mode=render
+    title: Tutorials
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/4zv8q/?direct%26mode=render%26action=download%26mode=render
+    title: Outro
+  - link: https://mfr.ca-1.osf.io/render?url=https://osf.io/bp4r6/?direct%26mode=render%26action=download%26mode=render
+    title: Reading
+  tutorials: 3
 - day: W2D2
   category: Dynamical Systems
   intro: https://www.youtube.com/watch?v=8NsLc0KoeYg
