### 2021-06-24 5c553c3
Add bonus autoencoder day to readme generation
