+++ tutorials/README.md
@@ -2,7 +2,7 @@
 
 <!-- DO NOT EDIT THIS FILE. IT IS AUTO-GENERATED BY A FRIENDLY ROBOT -->
 
-[W0D1](#w0d1---python-workshop1) | [W0D2](#w0d2---python-workshop2) | [W0D3](#w0d3---linear-algebra) | [W0D4](#w0d4---calculus) | [W0D5](#w0d5---statistics) | [W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---generalized-linear-models) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---biological-neuron-models) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
+[W0D1](#w0d1---python-workshop1) | [W0D2](#w0d2---python-workshop2) | [W0D3](#w0d3---linear-algebra) | [W0D4](#w0d4---calculus) | [W0D5](#w0d5---statistics) | [W1D1](#w1d1---model-types) | [W1D2](#w1d2---modeling-practice) | [W1D3](#w1d3---model-fitting) | [W1D4](#w1d4---generalized-linear-models) | [W1D5](#w1d5---dimensionality-reduction) | [W2D1](#w2d1---deep-learning) | [Bonus](#bonus---autoencoders) | [W2D2](#w2d2---linear-systems) | [W2D3](#w2d3---biological-neuron-models) | [W2D4](#w2d4---dynamic-networks) | [W3D1](#w3d1---bayesian-decisions) | [W3D2](#w3d2---hidden-dynamics) | [W3D3](#w3d3---optimal-control) | [W3D4](#w3d4---reinforcement-learning) | [W3D5](#w3d5---network-causality)
 
 *Warning:* The 'render with NBViewer' buttons may show outdated content.
 
@@ -188,6 +188,27 @@ Slides: [Intro](https://mfr.ca-1.osf.io/render?url=https://osf.io/jw829/?direct%
 
 
 [Further Reading](https://github.com/NeuromatchAcademy/course-content/blob/master/tutorials/W2D1_DeepLearning/further_reading.md)
+
+
+## Bonus - Autoencoders
+
+[YouTube Playlist](https://www.youtube.com/playlist?list=PLkBQOLLbi18Ojl1CV8W00JZ0C0hivBxw1)
+
+
+Slides: [Intro](https://mfr.ca-1.osf.io/render?url=https://osf.io/pdnsg/?direct%26mode=render%26action=download%26mode=render) | [Tutorials](https://mfr.ca-1.osf.io/render?url=https://osf.io/4quda/?direct%26mode=render%26action=download%26mode=render) | [Outro](https://mfr.ca-1.osf.io/render?url=https://osf.io/4zv8q/?direct%26mode=render%26action=download%26mode=render) | [Reading](https://mfr.ca-1.osf.io/render?url=https://osf.io/bp4r6/?direct%26mode=render%26action=download%26mode=render)
+
+|   | Run | View |
+| - | --- | ---- |
[…]
