+++ .github/workflows/notebook-pr.yaml
@@ -38,7 +38,6 @@ jobs:
           python -m pip install --upgrade pip wheel
           pip install -r requirements.txt
           pip install nbconvert pillow flake8 fuzzywuzzy[speedup]
-          pip install pyyaml
 
       - name: Install XKCD fonts
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
