+++ .github/workflows/tutorial-readme-pr.yaml
@@ -24,10 +24,6 @@ jobs:
         with:
           python-version: 3.7
 
-      - name: Install dependencies
-        run: |
-          pip install pyyaml
-
       - name: Update READMEs
         run: python ci/generate_tutorial_readmes.py
 
