### 2021-06-25 dab7d22
Delete generate_tutorial_readmes.py
