+++ ci/generate_tutorial_readmes.py
@@ -1,249 +1,0 @@
-"""Write a directory of tutorial notebooks to the README file.
-
-Run this script from the root of the github repository.
-
-"""
-import os
-from glob import glob
-import yaml
-
-def main():
-
-    # Initialize the lines in tutorials/README.md
-    course_readme_text = [
-    ]
-
-    try:
-        playlist_urls = load_youtube_playlist_urls()
-    except Exception as err:
-        print("Encountered error while loading youtube playlist links")
-        print(err)
-        playlist_urls = {}
-
-    try:
-        slide_urls = load_slide_urls()
-    except Exception as err:
-        print("Encountered error while loading slide links")
-        print(err)
-        slide_urls = {}
-
-    day_anchors = {}
-
-    day_paths = sorted(glob("tutorials/W?D?_*"))
-    for day_path in day_paths:
-
-        day_name = os.path.split(day_path)[-1]
-        day_code, topic_code = day_name.split("_")
-
-        # Split the UpperCamelCase topic name into separate words
-        topic_words = []
-        for letter in topic_code:
-            if letter.isupper():
-                topic_words.append(letter)
-            else:
-                topic_words[-1] += letter
-        topic = " ".join(topic_words)
-
-        # Note: this will fail if we have 10+ notebooks
-        notebooks = sorted(glob(f"{day_path}/*.ipynb"))
-
-        if not notebooks:
-            continue
-
-        # Track the anchor to this section for embed in the header
-        anchor = "-".join([
-            day_code.lower(),
-            "-",
-            ("-".join(topic_words)).lower(),
-        ])
-
-        day_anchors[day_code] = "#" + anchor
-
-        student_notebooks = get_student_links(notebooks)
-
-        # Write the day information into the course README
-        course_readme_text.extend([
-            f"## {day_code} - {topic}",
-            "",
-        ])
-
-        # Add a link to the YouTube lecture playlist, if we have one
-        youtube_url = playlist_urls.get(day_code, None)
-        if youtube_url is not None:
-            course_readme_text.extend([
-                f"[YouTube Playlist]({youtube_url})",
-                "",
-            ])
-
-        # Add links to slides
-        slide_links_by_topic = slide_urls.get(day_code, None)
-        if slide_links_by_topic is not None:
-            slide_links = [
-                f"[{topic}]({url})" for topic, url in slide_links_by_topic
-            ]
-
-            course_readme_text.extend([
-                "",
-                "Slides: " + " | ".join(slide_links),
-                "",
-            ])
-
-        # Add badges for all notebooks
-        course_readme_text.extend(write_badge_table(student_notebooks))
-        course_readme_text.append("\n")
-
-        # Add further reading
-        further_reading_file = f"{day_path}/further_reading.md"
-        if os.path.exists(further_reading_file):
-            course_readme_text.extend([f"[Further Reading](https://github.com/NeuromatchAcademy/course-content/blob/master/{further_reading_file})"])
-            course_readme_text.append("\n")
-
-        # Now make the day-specific README
-        # with links to both instructor and student versions
-        day_readme_text = [
-            f"# {day_code} - {topic}",
-            "",
-            "## Instructor notebooks",
-            "",
-        ]
-        day_readme_text.extend(write_badge_table(notebooks))
-
-        day_readme_text.extend([
-            "## Student notebooks",
-            "",
-        ])
-        day_readme_text.extend(write_badge_table(student_notebooks))
-
[…]
