### 2021-06-25 bd8e7fe
Install CI tools from centralized repo (#558)
* Install CI tools from centralized repo

* Add logic to install specific nmaci branch

* Instal the CI tools for the tutorial README PR too

* Prevent nmaci files from ending up in the sendback manifest

* Install CI deps from nmaci file

* Add execute flag

For compat with https://github.com/NeuromatchAcademy/nmaci/pull/1

* Add environment variable for https://github.com/NeuromatchAcademy/nmaci/pull/6
