+++ .github/workflows/notebook-pr.yaml
@@ -6,6 +6,7 @@ on:
 
 env:
   NB_KERNEL: python
+  NMA_MAIN_BRANCH: master
 
 jobs:
 
@@ -32,12 +33,22 @@ jobs:
         with:
           python-version: 3.7
 
+      - name: Install CI tools
+        if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
+        run: |
+          BRANCH=`python -c 'import os, re; m = re.search(r"nmaci:([\w-]+)", os.environ["COMMIT_MESSAGE"]); print("main" if m is None else m.group(1))'`
+          wget https://github.com/NeuromatchAcademy/nmaci/archive/refs/heads/$BRANCH.tar.gz
+          tar -xzf $BRANCH.tar.gz
+          pip install -r nmaci-$BRANCH/requirements.txt
+          mv nmaci-$BRANCH/scripts/ ci/
+          rm -r nmaci-$BRANCH
+          echo ci/ >> .gitignore
+
       - name: Install dependencies
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
         run: |
           python -m pip install --upgrade pip wheel
           pip install -r requirements.txt
-          pip install nbconvert pillow flake8 fuzzywuzzy[speedup]
 
       - name: Install XKCD fonts
         if: "!contains(env.COMMIT_MESSAGE, 'skip ci')"
@@ -60,7 +71,7 @@ jobs:
         run: |
           branch=${{ github.event.pull_request.head.ref }}
           nbs=`python ci/select_notebooks.py ${{ steps.changes.outputs.files }}`
-          python ci/process_notebooks.py $nbs
+          python ci/process_notebooks.py $nbs --execute
           python ci/verify_exercises.py $nbs --c "$COMMIT_MESSAGE"
           python ci/make_pr_comment.py $nbs --branch $branch --o comment.txt
 
