### 2021-06-25 44bbbbf
Clean up code
