### 2021-06-25 #118 @thartbm
Sorry, ramp-up for NMA 2021 is at full blast so everybody is distracted.
1. I agree that "NMA2020" would be a good version, but there are other considerations. First, I'd get the name out of the version, which would make it just "2020". Then I'd add that it is NMA about Computational Neuroscience (NMA-CN), since this year we also have a Deep Learning summer school (NMA-DL).
Preferred version: "2020.CN"
Runner up: "2020"
I think this can be set by whedon, or is that something we need to add to our course content GitHub repository?
2. I think I fixed the references.
3. The permanent link is supposed to be added after review is finished (if I read the instructions correctly). It would be good to start building that however. There is a zipped archive of the 2020 materials on the GitHub that we could use, but I'm not sure we want to leave it there (maybe a separate zenodo / OSF / other repository would be better) or still have to change the content. How we do this is not up to only me though, so this decision will unfortunately have to wait a bit.
