### 2021-06-25 #118 @thartbm
Updated / extended references look OK to me.
