### 2021-06-30 #118 @TomDonoghue
Hey all! Sorry my review is a bit slow, I had some unexpected work commitments the last couple weeks.

I think the NMA course and materials are fantastic, and reflect a wonderful collaborative project to help creating and offering quality educational resources. I don’t have any specific comments on the content, since NMA has done a fantastic job setting up an excellent pipeline for creating and validating materials, and I have nothing I can think to add to that process. I do have some potential suggestions for small updates to the repo description and the paper, that I think might help (but none I consider super critical, so I consider it broadly up to the authors to update as they wish).

### Version

As also mentioned by PaulScotti, there is something of an open question regarding the version that should be tagged in the paper. As I understand it, the paper is basically about the 2020 version, but there is about to be a new 2021 version. The paper could be stamped to the 2020 version, but it does seem a little weird to me that it would become immediately outdated in a couple weeks.

I think the notion of a JOSE paper is that it can refer to the materials, in general, and be a little future proofed to reflect the project as a whole. For example, if Table 1 noted that this is an “example topic list, from the 2020 version”, and maybe a single sentence was added to the end of the ‘Experience’ section that the plan is for the course to be taught annually, with corresponding updates to the materials, then it seems to me this paper would be more explicitly generalized across updates, and could be tagged to either the 2020 or 2021 versions.

### Repo Notes

- I think some quick additional notes in the README might help, in particular, for anyone looking into these materials from outside the context of being an NMA student. For example, it’s not entirely clear if/how you expect a student to self-start and/or if/how they should report any issue they find, etc.
	- For those looking to self-start, I think a couple overview sentences in the README might help to note that these materials are part of a course (link to website) but that for anyone looking to self-start, they can go to the notebooks (or JupyterBook site), with a quick note of what they would need to download and run locally - which is to install Python + install requirements, presumably?.  (This might also be described in an intro video, or page somewhere, in which case you could perhaps link it to the README?).
	- There is nothing clearly noted for how someone should interact with the repo, if they have a suggestion or notice an issue, etc. Since you have a whole procedure for material development, it’s not clear how an outsider should interact with that. Would it be worth adding a note in the README (and/or a CONTRIBUTING.md file) to briefly note if and how external people should file issues, open PRs, etc.
	- Relatedly (though optional), maybe it makes sense to add and mention a Code of Conduct?

### Paper Notes

- Software tools (suggest more explicitly noting / citing relevant software)
	- It’s sort of implicit (by mentioning Python introduction days) that the course is fully using Python. I think in the summary it could be more explicitly mentioned that this course uses Python, and maybe a quick note on the choice of this language.
	- Relatedly, there are a list of core dependencies in the requirements file (numpy, matplotlib, etc). I think it may be useful to mention the core dependencies (those in the requirements file, for example), both to clearly indicate to the audience what tooling they can expect to learn, and to cite and acknowledge the tools that this module is built on top of.
	- I think the pedagogical tools that are used to create these materials could be more explicitly mentioned / cited, for example Jupyter notebooks, with a link / citation to the Jupyter Project, to provide a reference for anyone who might be unfamiliar.
		- Relatedly, I think where “ipython” is currently noted, it should probably refer to “Jupyter” notebooks.
		- Though it depends a bit on how y’all decide to organize versions, I think JupyterBook could also be mentioned in passing, again to acknowledge the tools that are being used to support this module, noting that materials are now also hosted using JupyterBook, with a link to the site.
- Story of how this emerged
	- There is not a lot of description of how NMA evolved, and while the full story is probably out of scope of this paper, it seems relevant that “NeuroMatch” (as I understand it) is something of an umbrella name / organization, that includes the online conferences, and (upcomingly) other courses, such as the deep learning course. I think it might be worth briefly mentioning this in the paper, to clarify that there is a broader context within which this course exists, and that the conference was a precursor to this course. This can be very briefly described, and perhaps in part outsourced by citing / linking to the conference related paper (https://www.sciencedirect.com/science/article/pii/S1364661321000097).
- Formatting on Appendix 1 table
	- W1D1 & onwards, unclear gaps in the “Prerequisites column” -  a little hard to follow
	- W1D5 some overlap of rendering between “Topic” & “Prerequisites”
