### 2021-06-30 #118 @thartbm
These are all great points, thanks! I'll take it back to NMA and see what people think.
