### 2021-07-19 #118 @emckiernan
Hi @thartbm, just wanted to check in...any questions, or do you need anything on my end?
