### 2021-07-19 #118 @thartbm
@emckiernan Thanks for the reminder. Getting structured feedback from this large a group of authors is quite a bit of work, and since I'm overwhelmed with running NMA right now, I have unfortunately not been able to put in any work on this yet. It is absolutely on my priority list though!
