### 2021-10-05 #118 @thartbm
Thanks! We've wound down last summer's work, decided on the revisions and just started implementing them. After changes, I need to have it checked again (there are many, many people involved) but I think we should have an improved version relatively soon.

Can you help with how to set versions? Is that a Whedon command or something I need our GitHub people to set?
