### 2021-10-14 #118 @labarba
The DOI of this paper, when accepted and published, is already known. Your paper would resolve from `https://doi.org/10.21105/jose.00118` (the DOI is formed with the Review issue number).
