### 2022-02-26 NeuromatchAcademy/course-content#603 @thartbm
We now have an OSF repository with the GitHub content, as well as a JupyterBook version (doesn't have solutions, but nicer for students). Which is here: https://osf.io/93wna as well as a back up of the movies (in lower resolution) here: https://osf.io/79wv2/. I will need to be authorized to add a GitHub add-on to link the OSF repositories to the original repositories, but then they can be registered as permanent archives. Hope this solves it?
