### 2022-02-26 NeuromatchAcademy/course-content#602 @thartbm
I have added references to all data sets that have a citable reference (not all of them).
