### 2022-02-26 NeuromatchAcademy/course-content#601 @thartbm
The group has decided to make it about the 2021 version (with many other consequences) so I am now looking at how to get JOSE / Whedon to recognize a specific version.
