### 2022-03-13 NeuromatchAcademy/course-content#603 @thartbm
Turns out OSF has specific size restrictions, so the data sets had to be split into several repositories. But that seems all solved in this registered project: https://doi.org/10.17605/OSF.IO/CFN4U

The table of pre-requisites is now in the wiki page there (no longer in the paper), and there are 5 "links" to other registrations: 1 with videos of lectures, and 4 with data sets.
