### 2022-03-14 #118 @labarba
We just remove the `paused` label. I've done that now.
