### 2022-03-16 #118 @TomDonoghue
I've re-read the updated version of the paper, and had a quick look through the updates and materials, etc.

Overall, I think everything looks good to me!

I think a possible thing would be to still think about adding some notes to the repository on how to pursue self-study, how to report bugs, etc (see `Repo Notes` from my review from above), and now it might be useful to link the OSF links, etc.

However, this is a little tangential from the paper itself, so overall I'm happy to sign off on this and say I approve the revised paper.

Thanks to the authors and everyone involved for this rather immense / intense project! Congrats!

As a bit of a tangent, when I was exploring around, I noticed that the URL on the Github org page is wrong / doesn't resolve:
https://github.com/NeuromatchAcademy
