### 2022-03-16 #118 @thartbm
OK, I added a note about where to file bugs (issues are open to anyone on the content repo), and hopefully placed table 1 a bit better (the caption is now on a different page). I've told tech about the incorrect link on the org page. Not sure I can do much about guiding self-study more on the wiki at this point.
