### 2022-03-16 #118 @thartbm
@emckiernan Yes, please!

I think the last version of the pdf also looks better wrt to placement of table 1. Let me know what the next step is.
