### 2022-03-16 #118 @thartbm
@TomDonoghue, @PaulScotti Thanks for all the suggestions, it took some time, but IMO the paper got better because of your feedback.
