### 2022-03-16 #118 @emckiernan
Hi @thartbm, just a small correction/clarification. The link at the top of your GitHub repo resolves properly now, but it's different from the one listed in the paper Summary. The repo points to https://academy.neuromatch.io/, which does resolve, but the paper points to https://neuromatch.io/academy, which does not. Can you just confirm the link in the text should be the first one, and if so, could you please make that correction? Thanks!
