### 2022-03-16 #118 @thartbm
Good point. I guess they changed all the domains at some point. I can update the paper, but I'll also ask if we can stick to one plan for all this.
