### 2022-03-17 #118 @emckiernan
Hi @labarba, sorry to bug you! But we have a potential issue on the proofs for this paper. We've set the archive as  https://doi.org/10.17605/OSF.IO/CFN4U, but when we generate the pdf, the link under Software (pg. 3) doesn't work. (Review and Repository links work fine.) Am I missing anything?
