### 2022-03-17 #118 @labarba
@openjournals/dev 👋 Can you help us? The editor properly set the archive DOI, and the link at the top of this Review issue works, but for some reason the hyperlink in the PDF of the proof is not active (it seems not to have a target URL). Since this is done automatically on the paper compilation step, I'm not sure how we can troubleshoot this?
