### 2022-03-24 #118 @emckiernan
Hi @openjournals/dev, just wondering if you had an update on this? I'd like to recommend the paper for acceptance, but was hesitant to do so if there's still a bug to be worked out on the archive DOI. Thanks in advance for any help!
