### 2022-03-24 #118 @arfon
Apologies for the slow reply here. The archive DOI links only become active in the final proofs which are produced with the `recommend-accept` command.
