### 2022-03-24 #118 @labarba
Try compiling the paper here as a sanity check: https://whedon.theoj.org
