### 2022-03-24 #118 @emckiernan
Thanks, @labarba. I tried this and it failed with the following error: "Can't find any papers to compile. Make sure there's a file named paper.md in your repository."

@thartbm, it looks like there's no longer a file in your repo called paper.md.
