### 2022-03-24 #118 @emckiernan
> The repository given by the authors is https://github.com/NeuromatchAcademy/course-content/tree/jose_paper ... but we need https://github.com/NeuromatchAcademy/course-content/ ... to then proceed to compile from branch `jose_paper`.

I think this is a pathway issue. If we look here https://github.com/NeuromatchAcademy/course-content there's no paper.md, but here https://github.com/NeuromatchAcademy/course-content/tree/jose_paper there is. We had this same problem before, see the comment quoted above from @labarba.
