### 2022-03-24 #118 @thartbm
Would it help to put paper.md in the main branch? I can ask tech to move stuff around. The thing is that they don't want to "mess up" the main branch, but they now also want to close the `jose_paper` branch, because it is all getting rather big. Either way works for me... Would we remove the paper.md file at some point anyway? Or could we remove everything in the `jose_paper` branch except paper.md and the bibliography file?
