### 2022-03-25 #118 @emckiernan
> I just compiled the paper on the test service without problem, by adding the branch as shown on the screenshot

@labarba: Hmm, but then why didn't the command 'generate pdf from branch jose_paper' work? Doesn't it do the same thing?

Well, at least it works now, that's the important thing! Thank you!
