### 2022-03-25 #118 @labarba
> why didn't the command 'generate pdf from branch jose_paper' work?

I have no idea!

In any case, we will get the upgraded editorial bot soon, and one thing it does have is better handling of branches.
