### 2022-03-25 #118 @labarba
@thartbm
Pleas fix the DOI here: https://github.com/NeuromatchAcademy/course-content/blob/7d033193b94b9290f3e5cf09d61255e3d2f72279/bibliography.bib#L149
