### 2022-03-25 #118 @thartbm
The link to the DOI that I just corrected looks odd - although it works - maybe the proof didn't get updated? Other than that, it all looks great to me.
