### 2022-03-25 #118 @labarba
@arfon - you said [above](https://github.com/openjournals/jose-reviews/issues/118#issuecomment-1078343779):

> The archive DOI links only become active in the final proofs which are produced with the recommend-accept command.

We've run the `recommend-accept` command, and the archive link is not produced. There have been a few mysterious things happening on this thread, as we approach to publishing this paper. Is `@whedon` alright?
