### 2022-03-27 #118 @labarba
We're having issues again with a lazy editorial bot.

@openjournals/dev Any clue what could be happening here? We've had several instances of the bot not completing the compilation or getting stuck somewhere…
