### 2022-04-10 #118 @thartbm
Sorry to open this back up again. Many people were hoping to add this paper to their resume, but it's not showing up on everyone's Google Scholar or other profiles. It does for some, but for example not for me. Any idea what the issue is and if it can be solved?
