### 2022-04-10 #118 @labarba
you can just add the pub manually to your Google Scholar profile
